% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{above_cutoff}
\alias{above_cutoff}
\title{Return the indices of the rows that average above the cut off}
\usage{
above_cutoff(data, cutoff)
}
\description{
Args:
   @param data: Data to measure the average row and evaluate
                against the cutoff. Row = Genes, Col = Cells.
   @param cutoff: Threshold to be above to be kept.
}
\details{
Returns:
   @return: Returns a vector of row indicies to keep (are above the cutoff).
}

