% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annots}
\alias{annots}
\title{Generated classification for 10 normal cells and 10 tumor cells.}
\format{A data frame with 20 rows (cells) and 1 columns (classification)}
\usage{
annots
}
\description{
Generated classification for 10 normal cells and 10 tumor cells.
}
\keyword{datasets}
