% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_HMM.R
\name{assign_HMM_states_to_proxy_expr_vals}
\alias{assign_HMM_states_to_proxy_expr_vals}
\title{assign_HMM_states_to_proxy_expr_vals}
\usage{
assign_HMM_states_to_proxy_expr_vals(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{infercnv object}
}
\value{
infercnv_obj
}
\description{
Replaces the HMM state assignments with the cnv levels they represent.
}
