% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{average_over_ref}
\alias{average_over_ref}
\title{Remove the average of the genes of the reference observations from all 
observations' expression. Normalization by column.}
\usage{
average_over_ref(average_data, ref_observations, ref_groups)
}
\description{
Args:
   @param average_data: Matrix containing the data to remove average from
                        (this includes the reference observations).
                        Row = Genes, Col = Cells.
   @param ref_observations: Indices of reference observations.
                                  Only these are used in the average.
   @param ref_groups: A list of vectors of indices refering to the
                      different groups of the reference indices.
}
\details{
Returns:
   @return: Expression with the average gene expression in the reference 
           observations removed.
}

