% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\name{cellGene}
\alias{cellGene}
\title{Access the values for cellGene}
\usage{
cellGene(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
A list.
}
\description{
This function returns the list of values in cellGene
}
