% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{center_smoothed}
\alias{center_smoothed}
\title{Center data after smoothing. Center with in cells using median.}
\usage{
center_smoothed(data_smoothed)
}
\description{
Args:
   @param data_smoothed: Matrix to center.
                         Row = Genes, Col = cells.
}
\details{
Returns:
   @return: Matrix that is median centered.
            Row = Genes, Col = cells.
}

