% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{center_with_threshold}
\alias{center_with_threshold}
\title{Center data and threshold (both negative and postive values)}
\usage{
center_with_threshold(center_data, threshold)
}
\description{
Args:
   @param center_data: Matrix to center. Row = Genes, Col = Cells.
   @param threshold: Values will be required to be with -/+1 * 
                     threshold after centering.
Returns:
   @return: Centered and thresholded matrix
}

