% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{collapse_snp2gene}
\alias{collapse_snp2gene}
\title{collapse_snp2gene}
\usage{
collapse_snp2gene(
  infercnv_allele_obj,
  gene_annot,
  collapse_method = c("mean", "median"),
  cores = 5
)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele based obj}

\item{gene_annot}{a gene order annotation}

\item{collapse_method}{The method used to collapse snp into gene. default = median}

\item{cores}{A number of cores being used during parallel computing. default = 5}
}
\description{
This function aims to collapse infercnv_allele obj from snp level into gene level
}
