% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{collapse_snp2gene_highest}
\alias{collapse_snp2gene_highest}
\title{collapse_snp2gene_highest}
\usage{
collapse_snp2gene_highest(infercnv_allele_obj, gene_annot)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele based obj}

\item{gene_annot}{a gene order annotation}
}
\description{
This function aims to collapse infercnv_allele obj from snp level into gene level
by only leveraging the highest coverage snp with each gene
}
