% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{color.palette}
\alias{color.palette}
\title{Helper function allowing greater control over the steps in a color palette.
Source:http://menugget.blogspot.com/2011/11/define-color-steps-for-
              colorramppalette.html#more}
\usage{
color.palette(steps, between = NULL, ...)
}
\description{
Args:
   @param steps: Vector of colors to change use in the palette
   @param between: Steps where gradients change
}
\details{
Returns:
   @return: Color palette
}

