% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\name{Create_NGCHM}
\alias{Create_NGCHM}
\title{Create Next Generation Clustered Heat Map (NG-CHM)}
\usage{
Create_NGCHM(infercnv_obj, path_to_shaidyMapGen, out_dir, title = NULL,
  gene_symbol = NULL, x.center = NA, x.range = NA)
}
\arguments{
\item{infercnv_obj}{(S4) InferCNV S4 object holding expression data, gene location data, annotation information.}

\item{path_to_shaidyMapGen}{(string) Path to the java application ShaidyMapGen.jar}

\item{out_dir}{(string) Path to where the infercnv.ngchm output file should be saved to}

\item{title}{(string) Title that will be used for the heatmap}

\item{gene_symbol}{(string) Specify the label type that is given to the gene needed to create linkouts, default is NULL}

\item{x.center}{(integer) Center expression value for heatmap coloring.}

\item{x.range}{(integer) Values for minimum and maximum thresholds for heatmap coloring.}
}
\value{
Exports a NGCHM file named infercnv.ngchm and saves it to the output directory given to infercnv.
}
\description{
Create highly interactive heat maps for single cell expression data using 
Next Generation Clustered Heat Map (NG-CHM). NG-CHM was developed and 
maintained by MD Anderson Department of Bioinformatics and Computational 
Biology in collaboration with In Silico Solutions.
}
