% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{create_sep_list}
\alias{create_sep_list}
\title{Create a sepList forthe heatmap.3 plotting function given integer vectors
of rows and columns where speration should take place.
The expected input to the heatmap function is a list of 2 lists.
The first list are column based rectangles, and the second row.
To define a rectagle the index of the row or column where the line of the rectagle
should be placed is done with a vector of integers, left, bottom, right and top line.
Ie. list(list(c(1,0,3,10), c(5, 0, 10,10)), list(c(1,2,3,4)))}
\usage{
create_sep_list(row_count, col_count, row_seps = NULL, col_seps = NULL)
}
\description{
Args:
   @param row_count: Total number of rows
   @param col_count: Total number of columns
   @param row_seps: Vector of integers indices for row breaks
   @param col_seps: Vector of integer indices for column breaks
}
\details{
Returns
   @return: List of lists of vectors
}

