% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{CreateInfercnvObject}
\alias{CreateInfercnvObject}
\title{CreateInfercnvObject}
\usage{
CreateInfercnvObject(
  raw_counts_matrix,
  raw_allele_matrix = NULL,
  raw_coverage_matrix = NULL,
  gene_order_file,
  annotations_file,
  ref_group_names,
  delim = "\\t",
  max_cells_per_group = NULL,
  min_max_counts_per_cell = c(100, +Inf),
  snp_split_by = "::",
  chr_exclude = c("chrX", "chrY", "chrM")
)
}
\arguments{
\item{raw_counts_matrix}{the matrix of genes (rows) vs. cells (columns) containing the raw counts
If a filename is given, it'll be read via read.table()
otherwise, if matrix or Matrix, will use the data directly.}

\item{raw_allele_matrix}{the matrix of alternate allele,if provided, containing the raw counts 
If a filename is given, it'll be read via read.table()
otherwise, if matrix or Matrix, will use the data directly.}

\item{raw_coverage_matrix}{the matrix of coverage allele, if provided, containing the raw counts
If a filename is given, it'll be read via read.table()
otherwise, if matrix or Matrix, will use the data directly.}

\item{gene_order_file}{data file containing the positions of each gene along each chromosome in the genome.}

\item{annotations_file}{a description of the cells, indicating the cell type classifications}

\item{ref_group_names}{a vector containing the classifications of the reference (normal) cells to use for infering cnv}

\item{delim}{delimiter used in the input files}

\item{max_cells_per_group}{maximun number of cells to use per group. Default=NULL, using all cells defined in the annotations_file. This option is useful for randomly subsetting the existing data for a quicker preview run, such as using 50 cells per group instead of hundreds.}

\item{min_max_counts_per_cell}{minimum and maximum counts allowed per cell. Any cells outside this range will be removed from the counts matrix. default=(100, +Inf) and uses all cells. If used, should be set as c(min_counts, max_counts)}

\item{snp_split_by}{the string used for splitting the rowname of allele matrix (if allele data provided). default = "::"}

\item{chr_exclude}{list of chromosomes in the reference genome annotations that should be excluded from analysis.  Default = c('chrX', 'chrY', 'chrM')}
}
\value{
infercnv
}
\description{
Creation of an infercnv object. This requires the following inputs:
A more detailed description of each input is provided below:

The raw_counts_matrix:

           MGH36_P10_A01 MGH36_P10_E09 MGH36_P3_D05 MGH36_P4_F07 MGH36_P4_G01 ...
AL627309.1             0             0            0            1            0
CICP27                 0             0           16           71          101
AL627309.5             3            34           24           31           12
FO538757.1             0             0            2            0            3
AP006222.1             2             0            1           11            0
...

The raw_allele_matrix (if provided):

              MGH36_P10_A01 MGH36_P10_E09 MGH36_P3_D05 MGH36_P4_F07 MGH36_P4_G01 ...
chr1::1543953         0.001             0            0            0            0
chr1::1860776         0.000             2            9            0            0
chr1::7779884         0.001             4            0            0            0
chr1::7780866         0.001             4            0            0            0
chr1::7961718         0.000             0            0            3            0
...

The raw_coverage_matrix (if provided):

              MGH36_P10_A01 MGH36_P10_E09 MGH36_P3_D05 MGH36_P4_F07 MGH36_P4_G01 ...
chr1::1543953             1             0            0            0            0
chr1::1860776             0             6            9            0            0
chr1::7779884             2            12            0            0            0
chr1::7780866             2             9            0            0            0
chr1::7961718             0             0            0            3            0
...

The gene_order_file, contains chromosome, start, and stop position for each gene, tab-delimited:

            chr  start   stop
AL627309.1 chr1  89295 133723
CICP27     chr1 131025 134836
AL627309.5 chr1 141474 173862
FO538757.1 chr1 185217 195411
AP006222.1 chr1 257864 359681
AL732372.2 chr1 365389 522928
...

The annotations_file, containing the cell name and the cell type classification, tab-delimited.

                                V2    V3
MGH36_P10_A01 Microglia/Macrophage MGH36
MGH36_P10_E09 Microglia/Macrophage MGH36
MGH36_P3_D05  Microglia/Macrophage MGH36
MGH36_P4_F07             malignant MGH36
MGH36_P4_G01             malignant MGH36
MGH36_P4_H01             malignant MGH36
...


and the ref_group_names vector might look like so:  c("Microglia/Macrophage","Oligodendrocytes (non-malignant)")
}
\examples{
data(infercnv_data_example)
data(infercnv_annots_example)
data(infercnv_genes_example)
data(infercnv_allele_alt_example)
data(infercnv_allele_tot_example)

infercnv_object_example <- infercnv::CreateInfercnvObject(raw_counts_matrix = infercnv_data_example,
                                                          raw_allele_matrix = infercnv_allele_alt_example,
                                                          raw_coverage_matrix = infercnv_allele_tot_example,
                                                          gene_order_file = infercnv_genes_example,
                                                          annotations_file = infercnv_annots_example,
                                                          ref_group_names=c("Microglia/Macrophage"),
                                                          snp_split_by = "::")

}
