% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Generated SmartSeq2 expression data with 10 normal cells and 10 tumor cells.
This is only to demonstrate how to use methods, not actual data to be used in an analysis.}
\format{A data frame with 8252 rows (genes) and 20 columns (cells)}
\usage{
data
}
\description{
Generated SmartSeq2 expression data with 10 normal cells and 10 tumor cells.
This is only to demonstrate how to use methods, not actual data to be used in an analysis.
}
\keyword{datasets}
