% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_i3HMM.R
\name{determine_mean_delta_via_Z}
\alias{determine_mean_delta_via_Z}
\title{determine_mean_delta_via_Z}
\usage{
determine_mean_delta_via_Z(sigma, p)
}
\arguments{
\item{sigma}{standard deviation for a Normal distribution}

\item{p}{the p-value at which the distributions should intersect}
}
\value{
delta_for_alt_mean
}
\description{
determine means for amp/del distributions requiring that they cross the
             given distribution based on sigma centered at zero and at the given p value
}
