% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_i3HMM.R
\name{.i3HMM_get_normal_sd_trend_by_num_cells_fit}
\alias{.i3HMM_get_normal_sd_trend_by_num_cells_fit}
\title{.i3HMM_get_normal_sd_trend_by_num_cells_fit}
\usage{
.i3HMM_get_normal_sd_trend_by_num_cells_fit(infercnv_obj,
  i3_p_val = 0.05, plot = FALSE)
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{i3_p_val}{the p-value to use for defining the position of means for the alternate amp/del distributions.}

\item{plot}{boolean, set to TRUE to plot the mean/var fit.}
}
\value{
normal_sd_trend list
}
\description{
Determines the characteristics for the normal cell residual intensities, including
             fitting the variance in mean intensity as a function of number of cells sampled.
}
