% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\name{filterHighPNormals}
\alias{filterHighPNormals}
\title{filterHighPNormals: Filter the HMM identified CNV's by the CNV's posterior probability 
of belonging to a normal state.}
\usage{
filterHighPNormals(MCMC_inferCNV_obj)
}
\arguments{
\item{MCMC_inferCNV_obj}{MCMC infernCNV object.}
}
\value{
Returns a MCMC_inferCNV_obj With removed CNV's.
}
\description{
The following function will filter the HMM identified CNV's by the CNV's posterior 
probability of belonging to a normal state identified by the function inferCNVBayesNet(). Will filter 
CNV's based on a user desired threshold probability. Any CNV with a probability of being normal above 
the threshold will be removed.
}
