% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele_BayesNet.R
\name{filterHighPNormals_allele}
\alias{filterHighPNormals_allele}
\title{filterHighPNormals_allele: Filter the HMM (allele-based HMM or combined-based HMM) identified CNV's 
by the CNV's posterior probability of belonging to a normal state.}
\usage{
filterHighPNormals_allele(
  MCMC_inferCNV_obj,
  HMM_states,
  BayesMaxPNormal,
  reassignCNVs = T,
  HMM_type = c("i2", "i3", "i6")
)
}
\arguments{
\item{MCMC_inferCNV_obj}{MCMC infernCNV_allele or MCMC infernCNV_combined object.}

\item{HMM_states}{InferCNV object with HMM states in allele/both allele and expression data.}

\item{BayesMaxPNormal}{Option to filter CNV or cell lines by some probability threshold.}

\item{reassignCNVs}{(boolean) Given the CNV associated probability of belonging to each possible state, 
reassign the state assignments made by the HMM to the state that has the highest probability. (default: TRUE)}

\item{HMM_type}{The type of HMM that was ran, either 'i2' (allele), 'i3' (combined) or 'i6' (combined). 
Determines how many states were predicted by the HMM.}
}
\value{
Returns a list of (MCMC_inferCNV_allele_obj or MCMC_inferCNV_combined_obj, HMM_states) With removed CNV's.
}
\description{
The following function will filter the HMM identified CNV's by the CNV's posterior
probability of belonging to a normal state identified by the function inferCNVAlleleBayesNet() or
inferCNVCombinedBayesNet(). Will filter CNV's based on a user desired threshold probability. 
Any CNV with a probability of being normal above the threshold will be removed.
}
\examples{
data(infercnv_mcmc_obj_allele_gene_example)
data(infercnv_HMM_allele_gene_states_example)

mcmc_combined_mod_list <- filterHighPNormals_allele(MCMC_inferCNV_obj = mcmc_obj_allele_gene,
                                                    HMM_states = HMM_allele_gene_states,
                                                    BayesMaxPNormal = 0.5,
                                                    reassignCNVs = T,
                                                    HMM_type = "i2")
}
