% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_combined_BayesNet.R
\name{fusion_HMM_report_two_methods}
\alias{fusion_HMM_report_two_methods}
\title{fusion_HMM_report_two_methods}
\usage{
fusion_HMM_report_two_methods(
  expression_file_path,
  expression_file_token,
  allele_file_path,
  allele_file_token,
  infercnv_obj,
  infercnv_allele_obj,
  method = c("union", "common"),
  type = c("i6", "i3"),
  output_path,
  output_prefix,
  HMM_report_by = "subcluster",
  ...
)
}
\arguments{
\item{expression_file_path}{Location of the directory of the inferCNV outputs.}

\item{expression_file_token}{(string) String token that contains some info on settings used to name expression-based files.}

\item{allele_file_path}{Location of the directory of the inferCNV_allele outputs.}

\item{allele_file_token}{(string) String token that contains some info on settings used to name allele-based files.}

\item{infercnv_obj}{InferCNV object.}

\item{infercnv_allele_obj}{InferCNV allele object.}

\item{method}{The way integrating HMM boundaries from two methods: union/common.
"union" keeps those expression-specific gene that are not found in allele data during modeling,
while "common" removes those ones.}

\item{type}{The type of combined model running in Bayesian Method. i6 or i3.}

\item{output_path}{(string) Path to where the output file should be saved to.}

\item{output_prefix}{name prefix when naming output files.}

\item{HMM_report_by}{cell, consensus, subcluster (default: subcluster)}
}
\description{
This function aims to integrate HMM boundaries from expression/allele based methods.
}
