% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_HMM.R
\name{generate_cnv_region_reports}
\alias{generate_cnv_region_reports}
\title{generate_cnv_region_reports}
\usage{
generate_cnv_region_reports(infercnv_obj, output_filename_prefix, out_dir,
  ignore_neutral_state = NA, by = c("consensus", "subcluster", "cell"))
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{output_filename_prefix}{prefix for output filename}

\item{out_dir}{output directory for report files to be written}

\item{ignore_neutral_state}{numeric value representing the neutral state, which should be excluded from reporting (default: NA)}

\item{by}{options("consensus", "subcluster", "cell"), determines the granularity at which to report
the CNV regions.  Ideally, set to the same level at which the HMM predictions were performed.}
}
\value{
None
}
\description{
writes the CNV region report files
}
