% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{Downsampled gene coordinates file from GrCh37}
\format{A data frame with 10338 rows (genes) and 3 columns (chr, start, end)}
\usage{
genes
}
\description{
Downsampled gene coordinates file from GrCh37
}
\keyword{datasets}
