% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{get.sep}
\alias{get.sep}
\title{Get row or column lines of separation for heatmap.3}
\usage{
get.sep(clusters, type = c("row", "column", "both"))
}
\arguments{
\item{clusters}{a numerical vector, indicating the cluster labels of observations.}

\item{type}{string, one of the following: \code{c("row","column","both")}}
}
\description{
Please note this code is from the library GMD
All credit for this code goes to GMD's authors.
I do not recommend using this version of the code, which
has been poorly modified for our use but recommend using
the official version from the package GMD
https://cran.r-project.org/web/packages/GMD/index.html
Get row or column lines of separation for \code{heatmap.3} according to clusters
}

