% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_HMM.R
\name{get_hspike_cnv_mean_sd_trend_by_num_cells_fit}
\alias{get_hspike_cnv_mean_sd_trend_by_num_cells_fit}
\title{get_hspike_cnv_mean_sd_trend_by_num_cells_fit}
\usage{
get_hspike_cnv_mean_sd_trend_by_num_cells_fit(hspike_obj, plot = F)
}
\arguments{
\item{hspike_obj}{hidden spike object}

\item{plot}{(boolean flag, default FALSE)}
}
\value{
cnv_level_to_mean_sd_fit list
}
\description{
determine the number of cells - to - variance fit for each of the cnv levels.

Different numbers of cells are randomly selected from the distribution of residual intensitites at each
corresponding CNV level, the variance is computed, and a linear model is then fit.

Note, this is similar to what is done in HoneyBadger, but has many differences to how they're doing it there,
which appears to involve cnv block length rather than cell number.  Here, block size is not relevant, but rather
the number of cells in a pre-defined tumor subcluster.  Also, values are extracted from our in silico spike-in.
}
