% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_HMM.R
\name{get_predicted_CNV_regions}
\alias{get_predicted_CNV_regions}
\title{get_predicted_CNV_regions}
\usage{
get_predicted_CNV_regions(infercnv_obj, by = c("consensus", "subcluster",
  "cell"))
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{by}{options("consensus", "subcluster", "cell"), determines the granularity at which to report
the CNV regions.  Ideally, set to the same level at which the HMM predictions were performed.}
}
\value{
cnv_regions list
}
\description{
Given the infercnv_obj containing the HMM state assignments in the expr.data slot,
retrieves a list of CNV regions.
}
