% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_HMM.R
\name{get_spike_dists}
\alias{get_spike_dists}
\title{get_spike_dists}
\usage{
get_spike_dists(hspike_obj)
}
\arguments{
\item{hspike_obj}{hidden spike object}
}
\value{
cnv_mean_sd list
}
\description{
determines the N(mean,sd) parameters for each of the CNV states based on
the in silico spike in data (hspike).
}
