% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getCenter}
\alias{getCenter}
\alias{getCenter,NGCHM_inferCNV-method}
\title{Access the values for x.center}
\usage{
getCenter(obj)

\S4method{getCenter}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
x.center.
}
\description{
This function returns the list of values in x.center
}
