% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getChr}
\alias{getChr}
\alias{getChr,NGCHM_inferCNV-method}
\title{Get all the contigs in the correct order}
\usage{
getChr(obj)

\S4method{getChr}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Get all the contigs in the correct order
}
