% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getExpData}
\alias{getExpData}
\alias{getExpData,NGCHM_inferCNV-method}
\title{Access the values for exp.data}
\usage{
getExpData(obj)

\S4method{getExpData}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
exp.data
}
\description{
This function returns exp.data
}
