% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getGeneData}
\alias{getGeneData}
\alias{getGeneData,NGCHM_inferCNV-method}
\title{Access gene order from object}
\usage{
getGeneData(obj)

\S4method{getGeneData}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
gene_order.
}
\description{
This function returns the table of genes and their locations
}
