% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\docType{methods}
\name{getGenesCells}
\alias{getGenesCells}
\alias{getGenesCells,MCMC_inferCNV-method}
\title{Create a list that holds Genes and Cells for each separate identified CNV}
\usage{
getGenesCells(obj, pred_cnv_genes_df, cell_groups_df)

\S4method{getGenesCells}{MCMC_inferCNV}(obj, pred_cnv_genes_df,
  cell_groups_df)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Create a list that holds Genes and Cells for each separate identified CNV
}
