% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getObserved}
\alias{getObserved}
\alias{getObserved,NGCHM_inferCNV-method}
\title{Access the Observed indices}
\usage{
getObserved(obj)

\S4method{getObserved}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
list containing observed indices.
}
\description{
This function returns a list of the observed cell indices expression data
}
