% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{getProbabilities}
\alias{getProbabilities}
\alias{getProbabilities,MCMC_inferCNV-method}
\title{Set the probabilities for each CNV belonging to each state as well as probability of each cell belonging to a states}
\usage{
getProbabilities(obj)

\S4method{getProbabilities}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Set the probabilities for each CNV belonging to each state as well as probability of each cell belonging to a states
}
