% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getRowGrouping}
\alias{getRowGrouping}
\alias{getRowGrouping,NGCHM_inferCNV-method}
\title{Import the row groupings}
\usage{
getRowGrouping(obj)

\S4method{getRowGrouping}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{hm}{NGCHM object}
}
\value{
table containing the grouping of the rows
}
\description{
Import the row groupings
}
