% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\docType{methods}
\name{getStates}
\alias{getStates}
\alias{getStates,MCMC_inferCNV-method}
\title{Get the state values from the inferCNV HMM object}
\usage{
getStates(obj)

\S4method{getStates}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Get the state values from the inferCNV HMM object
}
