% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getTitle}
\alias{getTitle}
\alias{getTitle,NGCHM_inferCNV-method}
\title{Access the NGCHM title}
\usage{
getTitle(obj)

\S4method{getTitle}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
title.
}
\description{
This function returns the NGCHM title
}
