% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{getUniqueChr}
\alias{getUniqueChr}
\alias{getUniqueChr,NGCHM_inferCNV-method}
\title{Get the unique contigs in the correct order}
\usage{
getUniqueChr(obj)

\S4method{getUniqueChr}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Get the unique contigs in the correct order
}
