% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{heatmap.cnv}
\alias{heatmap.cnv}
\title{Please note this code is from the library GMD
All credit for this code goes to GMD's author's.
I do not recommend using this version of the code, which
has been poorly modified for our use but recommend using
the official version from the package GMD
https://cran.r-project.org/web/packages/GMD/index.html
This was originally heatmap.3.}
\usage{
heatmap.cnv(x, diss = inherits(x, "dist"), Rowv = TRUE, Colv = TRUE,
  dendrogram = c("both", "row", "column", "none"), dist.row, dist.col,
  dist.FUN = gdist, dist.FUN.MoreArgs = list(method = "euclidean"),
  hclust.row, hclust.col, hclust.FUN = hclust,
  hclust.FUN.MoreArgs = list(method = "ward"), scale = c("none", "row",
  "column"), na.rm = TRUE, cluster.by.row = FALSE, cluster.by.col = FALSE,
  kr = NA, kc = NA, row.clusters = NA, col.clusters = NA,
  revR = FALSE, revC = FALSE, add.expr, breaks, x.center,
  color.FUN = gplots::bluered, sepList = list(NULL, NULL),
  sep.color = c("gray45", "gray45"), sep.lty = 1, sep.lwd = 2, cellnote,
  cex.note = 1, notecol = "cyan", na.color = par("bg"),
  trace = c("none", "column", "row", "both"), tracecol = "cyan", hline,
  vline, linecol = tracecol, labRow = TRUE, labCol = TRUE,
  srtRow = NULL, srtCol = NULL, sideRow = 4, sideCol = 1,
  margin.for.labRow, margin.for.labCol, ColIndividualColors,
  RowIndividualColors, cexRow, cexCol, labRow.by.group = FALSE,
  labCol.by.group = FALSE, key = TRUE, key.title = "Color Key",
  key.xlab = "Value", key.ylab = "Count", keysize = 1.5, mapsize = 9,
  mapratio = 4/3, sidesize = 3, cex.key.main = 0.75,
  cex.key.xlab = 0.75, cex.key.ylab = 0.75, density.info = c("histogram",
  "density", "none"), denscol = tracecol, densadj = 0.25,
  main = "Heatmap", sub = "", xlab = "", ylab = "", cex.main = 2,
  cex.sub = 1.5, font.main = 2, font.sub = 3, adj.main = 0.5,
  mgp.main = c(1.5, 0.5, 0), mar.main = 3, mar.sub = 3, if.plot = TRUE,
  plot.row.partition = FALSE, plot.col.partition = FALSE,
  cex.partition = 1.25, color.partition.box = "gray45",
  color.partition.border = "#FFFFFF", plot.row.individuals = FALSE,
  plot.col.individuals = FALSE, plot.row.clusters = FALSE,
  plot.col.clusters = FALSE, plot.row.clustering = FALSE,
  plot.col.clustering = FALSE, plot.row.individuals.list = FALSE,
  plot.col.individuals.list = FALSE, plot.row.clusters.list = FALSE,
  plot.col.clusters.list = FALSE, plot.row.clustering.list = FALSE,
  plot.col.clustering.list = FALSE, row.data = FALSE, col.data = FALSE,
  if.plot.info = FALSE, text.box, cex.text = 1, force_lmat = NULL,
  force_lwid = NULL, force_lhei = NULL, force_add = FALSE, ...)
}
\description{
Please note this code is from the library GMD
All credit for this code goes to GMD's author's.
I do not recommend using this version of the code, which
has been poorly modified for our use but recommend using
the official version from the package GMD
https://cran.r-project.org/web/packages/GMD/index.html
This was originally heatmap.3.
}

