% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{highThreshold}
\alias{highThreshold}
\alias{highThreshold,NGCHM_inferCNV-method}
\title{Access the values for high_threshold}
\usage{
highThreshold(obj)

\S4method{highThreshold}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
high_threshold.
}
\description{
This function returns the list of values in high_threshold
}
