% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMM_allele_states}
\alias{HMM_allele_states}
\title{infercnv_allele object result of the processing of run() in the HMM example, to be used for other examples.}
\format{
An infercnv_allele object containing HMM predictions
}
\usage{
HMM_allele_states
}
\description{
infercnv_allele object result of the processing of run() in the HMM example, to be used for other examples.
}
\keyword{datasets}
