% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMM_obj}
\alias{HMM_obj}
\title{infercnv object result of the processing of run() in the HMM example, to be used for other examples.}
\format{An infercnv object containing HMM predictions}
\usage{
HMM_obj
}
\description{
infercnv object result of the processing of run() in the HMM example, to be used for other examples.
}
\keyword{datasets}
