% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_i3HMM.R
\name{i3HMM_assign_HMM_states_to_proxy_expr_vals}
\alias{i3HMM_assign_HMM_states_to_proxy_expr_vals}
\title{i3HMM_assign_HMM_states_to_proxy_expr_vals}
\usage{
i3HMM_assign_HMM_states_to_proxy_expr_vals(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{infercnv object}
}
\value{
infercnv_obj
}
\description{
replace i3 HMM state predictions with their represented CNV levels
}
