% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_i3HMM.R
\name{i3HMM_predict_CNV_via_HMM_on_tumor_subclusters}
\alias{i3HMM_predict_CNV_via_HMM_on_tumor_subclusters}
\title{i3HMM_predict_CNV_via_HMM_on_tumor_subclusters}
\usage{
i3HMM_predict_CNV_via_HMM_on_tumor_subclusters(infercnv_obj,
  i3_p_val = 0.05,
  normal_sd_trend = .i3HMM_get_normal_sd_trend_by_num_cells_fit(infercnv_obj,
  i3_p_val), t = 1e-06, use_KS = TRUE)
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{i3_p_val}{p-value used to determine mean for amp/del distributions}

\item{normal_sd_trend}{(optional) by default, computed automatically based on infercnv_obj, i3_p_val}

\item{t}{alt state transition probability (default: 1e-6)}

\item{use_KS}{boolean : use the KS test statistic to determine mean for amp/del dist HBadger style (default: TRUE)}
}
\value{
infercnv_obj where infercnv_obj@expr.data contains state assignments.
}
\description{
use the i3 HMM for predicting CNV at the level of tumor subclusters
}
