% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{infer_cnv}
\alias{infer_cnv}
\title{Infer CNV changes given a matrix of RNASeq counts. Output a pdf and matrix of final values.}
\usage{
infer_cnv(data, gene_order, cutoff, reference_obs, transform_data,
  window_length, max_centered_threshold, noise_threshold, num_ref_groups,
  out_path, plot_steps = FALSE, contig_tail = (window_length - 1)/2,
  method_bound_vis = NA, lower_bound_vis = NA, upper_bound_vis = NA)
}
\arguments{
\item{data:}{Expression matrix (genes X samples),
assumed to be log2(TPM+1) .}

\item{gene_order:}{Ordering of the genes (data's rows) 
according to their genomic location
To include all genes use 0.}

\item{cutoff:}{Cut-off for the average expression of genes to be 
used for CNV inference.}

\item{reference_obs:}{Column names of the subset of samples (data's columns)
that should be used as references.
If not given, the average of all samples will 
be the reference.}

\item{transform_data:}{Indicator to log2 + 1 transform}

\item{window_length:}{Length of the window for the moving average
(smoothing). Should be an odd integer.}

\item{max_centered_threshold:}{The maximum value a a value can have after
centering. Also sets a lower bound of
-1 * this value.}

\item{noise_threshold:}{The minimum difference a value can be from the 
average reference in order for it not to be
removed as noise.}

\item{num_ref_groups:}{The number of reference groups of a list of
indicies for each group of reference indices in
relation to reference_obs.}

\item{out_path:}{The path to what to save the pdf as. The raw data is
also written to this path but with the extension .txt .}

\item{plot_steps:}{If true turns on plotting intermediate steps.}

\item{contig_tail:}{Length of the tail removed from the ends of contigs.}

\item{method_bound:}{Method to use for bounding values in the visualization.}

\item{lower_bound_vis:}{Lower bound to normalize data to for visualization.}

\item{upper_bound_vis:}{Upper bound to normalize data to for visualization.}
}
\value{
Returns a list including:
    CNV matrix before visualization.
    CNV matrix after outlier removal for visualization.
    Contig order
    Column names of the subset of samples that should be used as references.
    Names of samples in reference groups.
}
