% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\docType{package}
\name{infercnv-package}
\alias{infercnv}
\alias{infercnv-package}
\title{infercnv: Infer Copy Number Variation from Single-Cell RNA-Seq Data}
\description{
Using single-cell RNA-Seq expression to visualize CNV in cells.
}
\details{
The main functions you will need to use are CreateInfercnvObject() and run(infercnv_object).
For additional details on running the analysis step by step, please refer to the example vignette.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/broadinstitute/inferCNV/issues}
}

}
\author{
\strong{Maintainer}: Timothy Tickle \email{ttickle@broadinstitute.org}

Authors:
\itemize{
  \item Itay Tirosh \email{tirosh@broadinstitute.org}
  \item Christophe Georgescu \email{cgeorges@broadinstitute.org}
  \item Maxwell Brown \email{mbrown@broadinstitute.org}
  \item Brian Haas \email{bhaas@broadinstitute.org}
}

}
