% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{infercnv}
\alias{infercnv}
\title{Run the infercnv analysis.}
\usage{
infercnv(x, gene_order, annotations = NULL, use_color_safe = FALSE,
  contig_label_size = 1, cutoff = 0, log_transform = FALSE,
  delim = "\\t", noise_filter = 0, max_centered_expression = 3,
  num_obs_groups = 1, output_dir = NULL, num_ref_groups = NULL,
  name_ref_groups = NULL, ref_subtract_method = "by_mean",
  hclust_method = "complete", clustering_contig = NULL,
  plot_steps = FALSE, bound_method_vis = "average_bound",
  bound_threshold_vis = " -1,1", window_length = 101, contig_tail = NULL,
  fig_title = "Copy Number Variation Inference",
  obs_title = "Observations (Cells)", ref_title = "References (Cells)",
  save_workspace = FALSE, log_level = "INFO")
}
\arguments{
\item{x}{Data to plot (columns are cells and rows are genes). Can be either a file path or the actual data frame.}

\item{gene_order}{Position information for each gene used to smooth over chromosomes. 
Can be either a file path or an actual data frame whose rows are genes and columns are the chromosome name, 
the start position, and the end position.}

\item{annotations}{A data frame with an annotation for each cell. It can be used with the name_ref_groups option 
to define which cells are to be used as references (and how they are grouped), 
and is used for a color bar on the side of the observations heatmap.}

\item{use_color_safe}{To support the needs of those who see colors differently, 
use this logical option to change the colors to a palette visibly distinct to all color blindness.}

\item{contig_label_size}{Used to increase or decrease the text labels for the X axis (contig names).}

\item{cutoff}{A number >= 0 is expected. A cut off for the average expression of genes to be used 
for CNV inference (use the value before log2 transformation).}

\item{log_transform}{Matrix is assumed to be Log2(TPM+1) transformed. 
If instead it is raw TPMs use this flag so that the data will be transformed.}

\item{delim}{Delimiter for reading expression matrix and writing matrices output.}

\item{noise_filter}{Delimiter for reading expression matrix and writing matrices output.}

\item{max_centered_expression}{This value and -1 * this value are used as the maximum value 
expression that can exist after centering data. If a value is outside of this range, 
it is truncated to be within this range}

\item{num_obs_groups}{Number of groups in which to break the observations.}

\item{output_dir}{Directory in which to save pdf and other output.}

\item{num_ref_groups}{Define a number of groups to make automatically 
by unsupervised clustering. This ignores annotations within references, 
but does not mix them with observations.}

\item{name_ref_groups}{Names of groups from the "annotations" table whose cells 
are to be used as reference groups.}

\item{ref_subtract_method}{Method used to subtract the reference values from the observations. 
Valid choices are: "by_mean", "by_quantiles".}

\item{hclust_method}{Method used for hierarchical clustering of cells. Valid choices are: 
"ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid".}

\item{clustering_contig}{When clustering observation samples, all genomic locations 
are used unless this option is given. The expected value is one of the contigs (Chr) 
in the genomic positions file (case senstive). All genomic positions will be plotted 
but only the given contig will be used in clustering / group creation.}

\item{plot_steps}{Using this argument turns on plotting intemediate steps. 
The plots will occur in the same directory as the output pdf. 
Please note this option increases the time needed to run}

\item{bound_method_vis}{Method to automatically detect and bound outliers. 
Used for visualizing. If both this argument and --vis_bound_threshold are given, 
this will not be used. Valid choice is : "average_bound".}

\item{bound_threshold_vis}{Used as upper and lower bounds for values 
in the visualization. If a value is outside this bound it will be replaced by the 
closest bound. Should be given in the form of 1,1 (upper bound, lower bound).}

\item{window_length}{Window length for the smoothing.}

\item{contig_tail}{Contig tail to be removed.}

\item{fig_title}{Plot title.}

\item{obs_title}{Title for the observations matrix.}

\item{ref_title}{Title for the reference matrix.}

\item{save_workspace}{Save workspace as infercnv.Rdata}

\item{log_level}{Logging level.}
}
\value{
No return, void.
}
\description{
Wrapper method that verifies settings and prepares the input for the other methods. 
Use this function unless you know what you are doing.
}
