% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\docType{class}
\name{infercnv_allele-class}
\alias{infercnv_allele-class}
\alias{infercnv_allele}
\title{The infercnv_allele class}
\description{
This class extends the functionality of infercnv class aiming to
incorporate the allele information.

Slots in the infercnv_allele object (besides infercnv) include:
}
\section{Slots}{

\describe{
\item{\code{allele.data}}{<matrix> the alternate allele data matrix.}

\item{\code{coverage.data}}{<matrix> the coverage allele data matrix.}

\item{\code{SNP_info}}{<GRanges> A GRanges object stores the SNP sites information given allele data.}
}}

