% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_allele_alt_example}
\alias{infercnv_allele_alt_example}
\title{Generated SmartSeq2 snp data (alternative) with 10 normal cells and 10 tumor cells.
This is only to demonstrate how to use methods, not actual data to be used in an analysis.}
\format{
A data frame with 97 rows (snps) and 20 columns (cells)
}
\usage{
infercnv_allele_alt_example
}
\description{
Generated SmartSeq2 snp data (alternative) with 10 normal cells and 10 tumor cells.
This is only to demonstrate how to use methods, not actual data to be used in an analysis.
}
\keyword{datasets}
