% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_annots_example}
\alias{infercnv_annots_example}
\title{Generated classification for 10 normal cells and 10 tumor cells.}
\format{
A data frame with 20 rows (cells) and 1 columns (classification)
}
\usage{
infercnv_annots_example
}
\description{
Generated classification for 10 normal cells and 10 tumor cells.
}
\keyword{datasets}
