% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_genes_example}
\alias{infercnv_genes_example}
\title{Downsampled gene coordinates file from GrCh37}
\format{
A data frame with 10338 rows (genes) and 3 columns (chr, start, end)
}
\usage{
infercnv_genes_example
}
\description{
Downsampled gene coordinates file from GrCh37
}
\keyword{datasets}
