% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_HMM_allele_gene_states_example}
\alias{infercnv_HMM_allele_gene_states_example}
\title{infercnv_allele object result of the processing of run() in the HMM example, to be used for other examples.}
\format{
An infercnv_allele object containing HMM predictions (gene level)
}
\usage{
infercnv_HMM_allele_gene_states_example
}
\description{
infercnv_allele object result of the processing of run() in the HMM example, to be used for other examples.
}
\keyword{datasets}
