% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_HMM_states_example}
\alias{infercnv_HMM_states_example}
\title{infercnv object result of the processing of run() in the HMM example, to be used for other examples.}
\format{
An infercnv object containing HMM predictions
}
\usage{
infercnv_HMM_states_example
}
\description{
infercnv object result of the processing of run() in the HMM example, to be used for other examples.
}
\keyword{datasets}
