% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_mcmc_obj_example}
\alias{infercnv_mcmc_obj_example}
\title{infercnv object result of the processing of inferCNVBayesNet in the example, to be used for other examples.}
\format{
An infercnv object containing posterior probability of CNV states
}
\usage{
infercnv_mcmc_obj_example
}
\description{
infercnv object result of the processing of inferCNVBayesNet in the example, to be used for other examples.
}
\keyword{datasets}
