% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_object_allele_example}
\alias{infercnv_object_allele_example}
\title{infercnv_allele object result of the processing of run() in the example, to be used for other examples.}
\format{
An infercnv_allele object
}
\usage{
infercnv_object_allele_example
}
\description{
infercnv_allele object result of the processing of run() in the example, to be used for other examples.
}
\keyword{datasets}
