% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infercnv_object_example}
\alias{infercnv_object_example}
\title{infercnv object result of the processing of run() in the example, to be used for other examples.}
\format{
An infercnv object
}
\usage{
infercnv_object_example
}
\description{
infercnv object result of the processing of run() in the example, to be used for other examples.
}
\keyword{datasets}
