% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele_BayesNet.R
\name{inferCNVAlleleBayesNet}
\alias{inferCNVAlleleBayesNet}
\title{inferCNVAlleleBayesNet: Run Bayesian Network Mixture Model Leveraging Allele Data 
To Obtain Posterior Probabilities For HMM Predicted States}
\usage{
inferCNVAlleleBayesNet(
  file_path,
  file_token,
  infercnv_allele_obj,
  allele_mode = c("snp_level", "gene_level"),
  output_path,
  cores = 5
)
}
\arguments{
\item{file_path}{Location of the directory of the inferCNV_allele outputs.}

\item{file_token}{(string) String token that contains some info on settings used to name allele-based files.}

\item{infercnv_allele_obj}{InferCNV_allele object.}

\item{allele_mode}{The type of allele data provided. "snp_level" or "gene_level".}

\item{output_path}{(string) Path to where the output file should be saved to.}

\item{cores}{Option to run parallel by specifying the number of cores to be used. (Default: 5)}
}
\value{
Returns a MCMC_inferCNV_allele_obj and posterior probability of being in one of two Copy Number Variation states
(i2 states: 1,Deletion; 2,Neutral) for CNV's identified by inferCNV_allele's HMM.
}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior
probability of being in one of two Copy Number Variation states 
(i2 states: 1,Deletion; 2,Neutral) for CNV's identified by inferCNV_allele's HMM. 
Posterior probabilities are found for the entire CNV cluster and each individual cell line in the CNV.
}
