% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_combined_BayesNet.R
\name{inferCNVCombinedBayesNet}
\alias{inferCNVCombinedBayesNet}
\title{inferCNVCombinedBayesNet: Run Bayesian Network Mixture Model Leveraging both
Expression and Allele Data To Obtain Posterior Probabilities For HMM Predicted States}
\usage{
inferCNVCombinedBayesNet(
  combined_file_path,
  combined_file_token,
  infercnv_obj,
  infercnv_allele_obj,
  allele_mode = c("snp_level", "gene_level"),
  type = c("i6", "i3"),
  enable_cnLOH = TRUE,
  output_path,
  cores = 5
)
}
\arguments{
\item{combined_file_path}{Location of the directory of the combined HMM outputs.}

\item{combined_file_token}{(string) String token that contains some info on settings used to name combined-based files.}

\item{infercnv_obj}{InferCNV object.}

\item{infercnv_allele_obj}{InferCNV allele object.}

\item{allele_mode}{The type of allele data provided. "snp_level" or "gene_level".}

\item{type}{The type of combined model running in Bayesian Method. i6 or i3.}

\item{enable_cnLOH}{Option for detecting cnLOH events by adding one more state in Bayesian Method. (Default: True)}

\item{output_path}{(string) Path to where the output file should be saved to.}

\item{cores}{Option to run parallel by specifying the number of cores to be used. (Default: 5)}
}
\value{
Returns a MCMC_infercnv_combined obj and posterior probability of being in one of three/six Copy Number Variation states
(i6 states: 1-2,Deletion; 3,Neutral; 4-6,Amplification; 7,cnLOH,if enable_cnLOH),
(i3 states: 1,Deletion; 2,Neutral; 3,Amplification; 4,cnLOH,if enable_cnLOH), for CNV's identified by inferCNV's HMM.
}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior
probability of being in one of three/six Copy Number Variation states 
(i6 states: 1-2,Deletion; 3,Neutral; 4-6,Amplification; 7,cnLOH,if enable_cnLOH),
(i3 states: 1,Deletion; 2,Neutral; 3,Amplification; 4,cnLOH,if enable_cnLOH),
for CNV's identified by inferCNV's HMM. Posterior probabilities are found for the entire CNV cluster 
and each individual cell line in the CNV.
}
