% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{initializeNGCHMObject}
\alias{initializeNGCHMObject}
\alias{initializeNGCHMObject,NGCHM_inferCNV-method}
\title{Initialize the NGCHM_inferCNV_obj object}
\usage{
initializeNGCHMObject(obj, args_parsed, infercnv_obj)

\S4method{initializeNGCHMObject}{NGCHM_inferCNV}(obj, args_parsed,
  infercnv_obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{args_parsed}{The arguments given to the function.}

\item{infercnv_obj}{InferCNV object.}
}
\value{
obj The NGCHM_inferCNV_obj S4 object.
}
\description{
Initialize the NGCHM_inferCNV_obj object
}
