% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{initializeObject}
\alias{initializeObject}
\alias{initializeObject,MCMC_inferCNV-method}
\title{Initialize the MCMC_inferCNV_obj object}
\usage{
initializeObject(obj, args_parsed, infercnv_obj, HMM_obj)

\S4method{initializeObject}{MCMC_inferCNV}(obj, args_parsed, infercnv_obj,
  HMM_obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Initialize the MCMC_inferCNV_obj object
}
