% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{invert_log2xplus1}
\alias{invert_log2xplus1}
\title{invert_log2xplus1()}
\usage{
invert_log2xplus1(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{infercnv_object}
}
\value{
infercnv_obj
}
\description{
Computes 2^x - 1
Updates infercnv_obj@expr.data
}
