% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{lowThreshold}
\alias{lowThreshold}
\alias{lowThreshold,NGCHM_inferCNV-method}
\title{Access the values for low_threshold}
\usage{
lowThreshold(obj)

\S4method{lowThreshold}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
low_threshold.
}
\description{
This function returns the list of values in low_threshold
}
