% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{make_ngchm}
\alias{make_ngchm}
\title{make_ngchm() : generates next gen heatmap}
\usage{
make_ngchm(infercnv_obj, out_dir = ".", title = "NGCHM",
  gene_symbol = NULL, path_to_shaidyMapGen = NULL)
}
\arguments{
\item{infercnv_obj}{An infercnv object}

\item{out_dir}{output directory (default: '.')}

\item{title}{title of the interactive heatmap (default: "NGCHM")}

\item{gene_symbol}{##TODO  (default: NULL)}

\item{path_to_shaidyMapGen}{path to the shaidyMapGen jar file (default: NULL)}
}
\description{
Function for Generating a next-generation heatmap
}
