% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{make_zero_NA}
\alias{make_zero_NA}
\title{make_zero_NA()

Converts zero to NA
Updates infercnv_obj@expr.data}
\usage{
make_zero_NA(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{}
}
\value{
infercnv_obj
}
