% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele_BayesNet.R
\docType{class}
\name{MCMC_infercnv_allele-class}
\alias{MCMC_infercnv_allele-class}
\alias{MCMC_infercnv_allele}
\title{MCMC infercnv_allele class}
\description{
This class extends the functionality of infercnv_allele class
aiming to coorporate MCMC object leveraging Bayesian model

Slots in the MCMC_infercnv_allele object (besides infercnv) include:
}
\section{Slots}{

\describe{
\item{\code{bugs_model}}{}

\item{\code{n.array}}{}

\item{\code{n.meta}}{}

\item{\code{alpha_prob}}{}
}}

