% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_combined_BayesNet.R
\docType{class}
\name{MCMC_infercnv_combined-class}
\alias{MCMC_infercnv_combined-class}
\alias{MCMC_infercnv_combined}
\title{MCMC infercnv_combined class}
\description{
This class extends the functionality of infercnv/infercnv_allele class
aiming to cooperate MCMC object leveraging Bayesian model

Slots in the MCMC_infercnv_allele object (besides infercnv) include:
}
\section{Slots}{

\describe{
\item{\code{bugs_model}}{a string path to where the bug file locates.}

\item{\code{cell_gene}}{cell_gene.}

\item{\code{cnv_regions}}{cnv_regions.}

\item{\code{infercnv_obj}}{gene_obj.}

\item{\code{infercnv_obj_mu}}{infercnv_obj_mu.}

\item{\code{infercnv_obj_sig}}{infercnv_obj_sig.}

\item{\code{infercnv_allele_obj}}{allele_obj.}
}}

