% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_obj_allele_gene}
\alias{mcmc_obj_allele_gene}
\title{infercnv_allele object result of the processing of inferCNVAlleleBayesNet in the example, to be used for other examples.}
\format{
An infercnv_allele object containing posterior probability of CNV states (gene level)
}
\usage{
mcmc_obj_allele_gene
}
\description{
infercnv_allele object result of the processing of inferCNVAlleleBayesNet in the example, to be used for other examples.
}
\keyword{datasets}
