% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{mcmcDiagnosticPlots}
\alias{mcmcDiagnosticPlots}
\alias{mcmcDiagnosticPlots,MCMC_inferCNV-method}
\title{Create Diagnostic Plots And Summaries.}
\usage{
mcmcDiagnosticPlots(obj)

\S4method{mcmcDiagnosticPlots}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Create Diagnostic Plots And Summaries in order to determine if convergence has occured.
}
