% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{MeanSD}
\alias{MeanSD}
\alias{MeanSD,MCMC_inferCNV-method}
\title{Get the cell Mean and Standard Deviation for identified cnv regions}
\usage{
MeanSD(obj)

\S4method{MeanSD}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Get the cell Mean and Standard Deviation for identified cnv regions
}
