% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{ngchm}
\alias{ngchm}
\title{ngchm() : generates next gen heatmap}
\usage{
ngchm(infercnv_obj, out_dir = ".", title = "NGCHM",
  gene_symbol = NULL, path_to_shaidyMapGen = NULL, x.range = NA,
  x.center = NA)
}
\arguments{
\item{infercnv_obj}{An infercnv object}

\item{out_dir}{output directory (default: '.')}

\item{title}{title of the interactive heatmap (default: "NGCHM")}

\item{gene_symbol}{##TODO  (default: NULL)}

\item{path_to_shaidyMapGen}{path to the shaidyMapGen jar file (default: NULL)}

\item{x.range}{(integer) Values for minimum and maximum thresholds for heatmap coloring.}

\item{x.center}{(integer) Center expression value for heatmap coloring.}
}
\description{
Function for Generating a next-generation heatmap
}
