% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{class}
\name{NGCHM_inferCNV-class}
\alias{NGCHM_inferCNV-class}
\alias{NGCHM_inferCNV}
\title{NGCHM_inferCNV class}
\value{
Returns a NGCHM_inferCNV_obj
}
\description{
Object that is used for the creation of a highly interactive heat maps for single cell 
expression data using Next Generation Clustered Heat Map (NG-CHM)
}
\section{Slots}{

\describe{
\item{\code{args}}{(list) The arguments given to the function.}

\item{\code{low_threshold}}{(integer) Values for minimum threshold for heatmap coloring.}

\item{\code{high_threshold}}{(integer) Values for maximum threshold for heatmap coloring.}

\item{\code{reference_cells}}{(character) Vector of reference cell ID's}

\item{\code{reference_groups}}{(character) Vector of reference cell group ID's}
}}

