% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{plot_cnv_observations}
\alias{plot_cnv_observations}
\title{Plot the observational samples}
\usage{
plot_cnv_observations(obs_data, col_pal, contig_colors, contig_labels,
  contig_names, contig_seps, num_obs_groups, file_base_name, cnv_title,
  cnv_obs_title, contig_lab_size = 1, cluster_contig = NULL,
  testing = FALSE, layout_lmat = NULL, layout_lhei = NULL,
  layout_lwid = NULL)
}
\description{
Args:
   @param obs_data: Data to plot as observations. Rows = Cells, Col = Genes
   @param col_pal: The color palette to use.
   @param contig_colors: The colors for the contig bar.
   @param contig_labels: The labels for the contigs.
   @param contig_names: Names of the contigs
   @param contig_seps: Indices for line seperators of contigs.
   @param num_obs_groups: Number of groups of observations to create
   @param file_base_name: Base of the file to used to make output file names.
   @param cnv_title: Title of the plot.
   @param cnv_obs_title: Title for the observation matrix.
   @param contig_lab_size: Text size for contigs.
   @param cluster_contig: A value directs cluster to only genes on this contig
   @param layout_lmat: lmat values to use in layout
   @param layout_lhei: lhei values to use in layout
   @param layout_lwid: lwid values to use in layout
}
\details{
Returns:
   @return: Void
}

