% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{plot_cnv_references}
\alias{plot_cnv_references}
\title{Plot the reference samples}
\usage{
plot_cnv_references(ref_data, ref_groups, col_pal, contig_seps, file_base_name,
  cnv_ref_title, layout_lmat = NULL, layout_lwid = NULL,
  layout_lhei = NULL, layout_add = FALSE, testing = FALSE)
}
\description{
Args:
   @param ref_data: Data to plot as references. Rows = Cells, Col = Genes
   @param ref_groups: Groups of references to plot together.
   @param col_pal: The color palette to use.
   @param contig_seps: Indices for line seperators of contigs.
   @param file_base_name: Base of the file to used to make output file names.
   @param cnv_ref_title: Title for reference matrix.
   @param layout_lmat: lmat values to use in the layout.
   @param layout_lwid: lwid values to use in the layout.
   @param layout_lhei: lhei values to use in the layout.
   @param layout_add: Indicates the ref image shoudl be added to the previous plot.
   @param testing: Turns off plotting when true.
}
\details{
Returns:
   @return: Void
}

