% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{plot_observations_layout}
\alias{plot_observations_layout}
\title{Create the layout for the plot
This is a modification of the original
layout from the GMD heatmap.3 function}
\usage{
plot_observations_layout()
}
\description{
Returns:
   @return: list with slots "lmat" (layout matrix),
                            "lhei" (height, numerix vector),
                            and "lwid" (widths, numeric vector)
}

