% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{plot_step}
\alias{plot_step}
\title{Log intermediate step with a plot and text file of the steps.}
\usage{
plot_step(data, plot_name)
}
\description{
Args:
   @param data: The data frame to plot.
   @param plot_name: The absolute path to the pdf to be plotted.
}
\details{
Returns:
   @return: No return
}

