% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\docType{methods}
\name{plotProbabilities}
\alias{plotProbabilities}
\alias{plotProbabilities,MCMC_inferCNV-method}
\title{Main plotting function, plots the probability for each cnv belonging to a specific state and the probability of 
each cell line belonging to a specific states.}
\usage{
plotProbabilities(obj)

\S4method{plotProbabilities}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Main plotting function, plots the probability for each cnv belonging to a specific state and the probability of 
each cell line belonging to a specific states.
}
