% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_HMM.R
\name{predict_CNV_via_HMM_on_indiv_cells}
\alias{predict_CNV_via_HMM_on_indiv_cells}
\title{predict_CNV_via_HMM_on_indiv_cells}
\usage{
predict_CNV_via_HMM_on_indiv_cells(infercnv_obj,
  cnv_mean_sd = get_spike_dists(infercnv_obj@.hspike), t = 1e-06)
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{cnv_mean_sd}{(optional, by default automatically computed based in the infercnv_obj@.hspike object)}

\item{t}{HMM alt state transition probability (default=1e-6)}
}
\value{
infercnv_obj where the infercnv_obj@expr.data are replaced with the HMM state assignments.
}
\description{
predict CNV levels at the individual cell level, using the i6 HMM
}
