% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{remove_noise}
\alias{remove_noise}
\title{Remove values that are too close to the average and are considered noise.}
\usage{
remove_noise(smooth_matrix, threshold)
}
\description{
Args:
   @param smooth_matrix: A matrix of values, smoothed, and with average 
                         reference removed. Row = Genes, Col = Cells.
   @param threshold: The amount of difference a value must be from the
                     reference before the value can be kept and not 
                     removed as noise.
Returns:
   @return: Denoised matrix
}

