% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{remove_outliers_norm}
\alias{remove_outliers_norm}
\title{Set outliers to some upper or lower bound. Then normalize values to
approximately [-1, 1]. This is to prep the data for visualization.}
\usage{
remove_outliers_norm(data, out_method = NA, lower_bound = NA,
  upper_bound = NA, plot_step = NA)
}
\description{
Args:
   @param data: data to remove outliers. Outliers removed within columns.
   @param out_method: Method to remove outliers [(average_bound, NA (hard threshold))]
   @param lower_bound: Lower bound which identifies a measurement
                       as an outlier. 
   @param upper_bound: Upper bound which identifies a measurement
                       as an outlier.
   @param plot_step: True will plot this analysis step.
}
\details{
Returns:
   @return: Return data matrix with outliers removed
}

