% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_spike.R
\name{remove_spike}
\alias{remove_spike}
\title{remove_spike()}
\usage{
remove_spike(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{An infercnv object populated with raw count data}
}
\value{
infercnv_obj
}
\description{
Removes the spiked-in group named 'SPIKE' from the infercnv_obj
}
