% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{remove_tails}
\alias{remove_tails}
\title{Remove the tails of values of a specific chromosome.
The smooth_matrix values are expected to be in genomic order.
If the tail is too large and no contig will be left 1/3 of the
contig is left.}
\usage{
remove_tails(smooth_matrix, chr, tail_length)
}
\description{
Args:
   @param smooth_matrix: Smoothed values in genomic order.
                         Row = Genes, Col = Cells.
   @param chr: Indices of the chr in which the tails are to be removed.
   @param tail_length: Length of the tail to remove on both ends of the
                       chr indices.
Returns:
   @return: Indices to remove.
}

