% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\docType{methods}
\name{removeCells}
\alias{removeCells}
\alias{removeCells,MCMC_inferCNV-method}
\title{Run simulations and remove cells from cnv's that are predicted to be normal}
\usage{
removeCells(obj)

\S4method{removeCells}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Run simulations and remove cells from cnv's that are predicted to be normal
}
