% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\docType{methods}
\name{removeCNV}
\alias{removeCNV}
\alias{removeCNV,MCMC_inferCNV-method}
\title{Run simulations and remove cnv's that are predicted to be normal}
\usage{
removeCNV(obj)

\S4method{removeCNV}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Run simulations and remove cnv's that are predicted to be normal
}
