% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{require_above_min_cells_ref}
\alias{require_above_min_cells_ref}
\title{require_above_min_cells_ref()}
\usage{
require_above_min_cells_ref(infercnv_obj, min_cells_per_gene)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{min_cells_per_gene}{int indicating number of cells required per gene for both obs and ref data}
}
\value{
infercnv_obj infercnv_object with corresponding genes removed.
}
\description{
Filters out genes that have fewer than specified number of cells expressing them.
}
