% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{require_above_min_mean_expr_cutoff}
\alias{require_above_min_mean_expr_cutoff}
\title{require_above_min_mean_expr_cutoff ()

Filters out genes that have fewer than the corresponding mean value across the reference cell values.}
\usage{
require_above_min_mean_expr_cutoff(infercnv_obj, min_mean_expr_cutoff)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{min_mean_expr_cutoff}{the minimum mean value allowed for a gene to be retained in the expression matrix.}
}
\value{
infercnv_obj  the infercnv_object with lowly or unexpressed genes removed.
}
