% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\docType{methods}
\name{runMCMC}
\alias{runMCMC}
\alias{runMCMC,MCMC_inferCNV-method}
\title{Run simulations using rjags.}
\usage{
runMCMC(obj)

\S4method{runMCMC}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Run simulations using rjags.
}
