% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_spike.R
\name{scale_cnv_by_spike}
\alias{scale_cnv_by_spike}
\title{scale_cnv_by_spike()}
\usage{
scale_cnv_by_spike(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{An infercnv object populated with raw count data}
}
\value{
infercnv_obj
}
\description{
Scales expression data according to the expression value bounds in the SPIKE group.
}
\details{
Assumes data is centered at 1
Expression below 1 is scaled according to the left spike bound set to zero.
Expression above 1 is scaled according to the right spike bound set to two.
}
