% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{scale_infercnv_expr}
\alias{scale_infercnv_expr}
\title{scale_infercnv_expr}
\usage{
scale_infercnv_expr(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{infercnv object}
}
\value{
infercnv_obj
}
\description{
performs scaling to expression values for each cell,
assigning all values to a standard normal centered at zero.
}
