% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{setAlleleMatrix}
\alias{setAlleleMatrix}
\title{setAlleleMatrix}
\usage{
setAlleleMatrix(
  infercnv_allele_obj,
  snp_filter = TRUE,
  snp_het.deviance.threshold = 0.1,
  snp_min.cell = 3,
  smooth_window = 101
)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele_object}

\item{snp_filter}{a boolean value whether to do pre-filtering for allele matrix. default = TRUE}

\item{snp_het.deviance.threshold}{a threshold used to filter 
out non-hetergozous snps. default = 0.1}

\item{snp_min.cell}{a threshold used to filter out snps that express less than 
the minimum number of cells. default = 3}

\item{smooth_window}{a window size used for smoothing allele fraction. default = 101}
}
\description{
This function aims to set lesser allele fraction in-silico way.
lesser allele fraction will be stored in the slot @expr.data. lesser allele count 
will be stored in the slot @count.data
}
