% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{setAlleleMatrix_HB}
\alias{setAlleleMatrix_HB}
\title{setAlleleMatrix_HB -- deprecated}
\usage{
setAlleleMatrix_HB(
  infercnv_allele_obj,
  snp_min_coverage = 0,
  snp_filter = TRUE,
  snp_het.deviance.threshold = 0.1,
  snp_min.cell = 3
)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele_object}

\item{snp_min_coverage}{the minimum number of threshold that each instance should have in allele/coverage data. 
Each instance that is less than this number will be replaced with 0. default = 0 (No action)}

\item{snp_filter}{a boolean value whether to do pre-filtering for allele matrix. default = TRUE}

\item{snp_het.deviance.threshold}{a threshold used to filter 
out non-hetergozous snps. default = 0.1}

\item{snp_min.cell}{a threshold used to filter out snps that express less than 
the minimum number of cells. default = 3}
}
\value{
infercnv_allele
}
\description{
This function mimics the way the HoneyBadger does aiming to initialize the lesser allele fraction/count.
lesser allele fraction will be stored in the slot @expr.data. lesser allele count 
will be stored in the slot @count.data
}
