% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\name{setBayesMaxPNormal}
\alias{setBayesMaxPNormal}
\title{Add the probability threshold for the arguments in the MCMC infercnv object.}
\usage{
setBayesMaxPNormal(obj, BayesMaxPNormal)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}

\item{BayesMaxPNormal}{probability to be used as a threshold for CNV or cell removal.}
}
\value{
MCMC_inferCNV_obj S4 object.
}
\description{
This function adds the variable BayesMaxPNormal to the arguments slot of the the MCMC infercnv object.
}
