% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setColCovariateBar}
\alias{setColCovariateBar}
\alias{setColCovariateBar,NGCHM_inferCNV-method}
\title{Set Covariate bar for reference and observed groups}
\usage{
setColCovariateBar(obj, hm)

\S4method{setColCovariateBar}{NGCHM_inferCNV}(obj, hm)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Set Covariate bar for reference and observed groups
}
