% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setColors}
\alias{setColors}
\alias{setColors,NGCHM_inferCNV-method}
\title{Set color map for heatmap}
\usage{
setColors(obj)

\S4method{setColors}{NGCHM_inferCNV}(obj)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}
}
\value{
layer: color mapping object used for NGCHM heatmap generation
}
\description{
Set color map for heatmap
}
