% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setCutsColor}
\alias{setCutsColor}
\alias{setCutsColor,ngchmVersion2-method}
\title{Set the initial coloring for the cuts (contig and observed/reference divisions)}
\usage{
setCutsColor(hm)

\S4method{setCutsColor}{ngchmVersion2}(hm)
}
\arguments{
\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Set the initial coloring for the cuts (contig and observed/reference divisions)
}
