% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setDivisions}
\alias{setDivisions}
\alias{setDivisions,NGCHM_inferCNV-method}
\title{Set the divisions in the heatmap}
\usage{
setDivisions(obj, hm)

\S4method{setDivisions}{NGCHM_inferCNV}(obj, hm)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Set the divisions in the heatmap
}
