% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setHeatMapSize}
\alias{setHeatMapSize}
\alias{setHeatMapSize,ngchmVersion2-method}
\title{Option to adjust the size of the heat map.}
\usage{
setHeatMapSize(hm, width, hight)

\S4method{setHeatMapSize}{ngchmVersion2}(hm, width, hight)
}
\arguments{
\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Option to adjust the size of the heat map.
}
