% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setLabelSize}
\alias{setLabelSize}
\alias{setLabelSize,ngchmVersion2-method}
\title{Option to set the label display size}
\usage{
setLabelSize(hm, size)

\S4method{setLabelSize}{ngchmVersion2}(hm, size)
}
\arguments{
\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Option to set the label display size
}
