% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setNGCHMObject}
\alias{setNGCHMObject}
\alias{setNGCHMObject,NGCHM_inferCNV-method}
\title{Initialize the NGCHM object}
\usage{
setNGCHMObject(obj, hm)

\S4method{setNGCHMObject}{NGCHM_inferCNV}(obj, hm)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Initialize the NGCHM object
}
