% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\docType{methods}
\name{setRows}
\alias{setRows}
\alias{setRows,NGCHM_inferCNV-method}
\title{Set the Dendrograms and rows}
\usage{
setRows(obj, hm)

\S4method{setRows}{NGCHM_inferCNV}(obj, hm)
}
\arguments{
\item{obj}{The NGCHM_inferCNV_obj S4 object.}

\item{hm}{NGCHM object}
}
\value{
hm: The NGCHM S4 object.
}
\description{
Set the Dendrograms and rows
}
