% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{smooth_by_chromosome}
\alias{smooth_by_chromosome}
\title{smooth_by_chromosome()}
\usage{
smooth_by_chromosome(infercnv_obj, window_length, smooth_ends = TRUE)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{window_length}{length of window (number of genes) for the moving average}

\item{smooth_ends}{perform smoothing at the ends of the chromosomes (default: TRUE)}
}
\value{
infercnv_obj
}
\description{
Smooth expression values for each cell across each chromosome by using a
moving average with a window of specified length.
}
