% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{smooth_by_chromosome_runmeans}
\alias{smooth_by_chromosome_runmeans}
\title{smooth_by_chromosome_runmeans}
\usage{
smooth_by_chromosome_runmeans(infercnv_obj, window_length)
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{window_length}{window length to use for smoothing.}
}
\value{
infercnv_obj
}
\description{
uses the simpler caTools:runmeans() to perform smoothing operations.
}
