% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{smooth_ends_helper}
\alias{smooth_ends_helper}
\title{Helper function for smoothing the ends of a moving average.}
\usage{
smooth_ends_helper(obs_data, obs_tails)
}
\description{
Args:
   @param obs_data: Data to smooth
   @param obs_tails: Length of the tail to smooth.
}
\details{
Returns:
   @return: Data smoothed.
}

