% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{smooth_window}
\alias{smooth_window}
\title{Smooth a matrix by column using a simple moving average.
Tails of the averages use a window length that is truncated to
available data.}
\usage{
smooth_window(data, window_length)
}
\description{
Args:
   @param data: Data matrix to smooth. Row = Genes, Col = Cells.
   @param window_length: Length of window to use for the moving average.
       Should be a positive, odd integer.
}
\details{
Returns:
   @return: Matrix with columns smoothed with a simple moving average.
}

