% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{smooth_window_helper}
\alias{smooth_window_helper}
\title{Smooth vector of values over the given window length.}
\usage{
smooth_window_helper(obs_data, window_length)
}
\description{
Args:
   @param obs_data: Vector of data to smooth with a moving average.
   @param window_length: Length of the window for smoothing.
       Must be and odd, positive, integer.
}
\details{
Returns:
   @return: Vector of values smoothed with a moving average.
}

