% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_spike.R
\name{spike_in_variation_chrs}
\alias{spike_in_variation_chrs}
\title{spike_in_variation_chrs()

Adds a 'SPIKE'-in to the observations set at different thresholds of loss/gain to
aid in tracking the effect of infercnv operations and for defining the final scaling.}
\usage{
spike_in_variation_chrs(infercnv_obj, spike_in_chrs = NULL,
  spike_in_multiplier_vec = c(0.01, 2), max_cells = 100,
  min_genes_per_chr = 100)
}
\arguments{
\item{infercnv_obj}{An infercnv object populated with raw count data}

\item{spike_in_chrs}{: define the chromsomes that will serve as signal for gain/loss
default: picks chrosomes in order of size}

\item{spike_in_multiplier_vec}{: factors that define relative expression for gain/loss
and must match ordering of spike_in_chrs above
default: c(0.01, 2.0)}

\item{max_cells}{max number of cells to incorporate in the spike-in}

\item{min_genes_per_chr}{: default 100}
}
