% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{split_references}
\alias{split_references}
\title{Split up reference observations in to k groups and return indices
for the different groups.}
\usage{
split_references(average_data, ref_obs, num_groups)
}
\description{
Args:
   @param average_data: Matrix containing data. Row = Genes, Col = Cells.
   @param ref_obs: Indices of reference obervations.
   @param num_groups: The number of groups to partition nodes in or a list
                      of already partitioned indices.
}
\details{
Returns:
   @return: Returns a list of grouped reference observations given as
            vectors of groups. These are indices relative to the reference
            observations only, so a return 1 indicates the first reference
            row, not the first row.
}

