% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_mask_non_DE.R
\name{subcluster_tumors}
\alias{subcluster_tumors}
\title{subcluster_tumors

Divides tumor groupings according to clades by cutting the hierarchical cluster tree at specified height.}
\usage{
subcluster_tumors(infercnv_obj,
  tumor_groupings = infercnv_obj@observation_grouped_cell_indices,
  cut_tree_height_ratio = 0.9, hclust_method = "ward.D",
  min_median_tree_height_ratio = 2.5)
}
\arguments{
\item{infercnv_obj}{}

\item{tumor_groupings}{list of tumor group indices (default: infercnv_obj@observation_grouped_cell_indices)}

\item{cut_tree_height_ratio}{ratio of the hierarchical cluster tree height for cutting into subclusters (default: 0.9)}

\item{hclust_method}{see 'hclust' documentation. default="ward.D"}
}
\value{
list of tumor subclusters (ie.  list[[ tumor_subcluster_name ]] = c(cell_idxA, ..., cell_idxN)
}
\description{
subcluster_tumors

Divides tumor groupings according to clades by cutting the hierarchical cluster tree at specified height.
}
