% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{subtract_ref_expr_from_obs}
\alias{subtract_ref_expr_from_obs}
\title{subtract_ref_expr_from_obs()}
\usage{
subtract_ref_expr_from_obs(infercnv_obj, inv_log = FALSE)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{inv_log}{invert log-transformed data before subtracting mean, then convert back to log space.}
}
\value{
infercnv_obj containing the reference subtracted values.
}
\description{
Remove the average of the genes of the reference observations from all
observations' expression. In the case there are multiple reference groupings,
the averages are computed separately for each reference grouping, and the min|max
of the averages are subtracted from the observation expression levels. Any values within the range
of the min,max of the group are set to zero.
}
\details{
Subtracting the mean of the reference expr distributions from the observed cells.
}
