% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{transform_to_reference_based_Zscores}
\alias{transform_to_reference_based_Zscores}
\title{transform_to_reference_based_Zscores()}
\usage{
transform_to_reference_based_Zscores(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{infercnv_object}
}
\value{
infercnv_obj
}
\description{
Computes mean and standard deviation for the reference cells, then uses these values
to compute gene Z-scores for both the reference and observation cells.

Note, reference cell gene expression values will be centered at zero after this operation.
}
