% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{withParallel}
\alias{withParallel}
\alias{withParallel,MCMC_inferCNV-method}
\title{Run simulations in Parallel}
\usage{
withParallel(obj)

\S4method{withParallel}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Run simulations in Parallel
}
