% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_interaction.R
\name{add_to_seurat}
\alias{add_to_seurat}
\title{add_to_seurat()}
\usage{
add_to_seurat(seurat_obj = NULL, infercnv_output_path, top_n = 10,
  bp_tolerance = 2e+06)
}
\arguments{
\item{seurat_obj}{Seurat object to add meta.data to (default: NULL)}

\item{infercnv_output_path}{Path to the output folder of the infercnv run to use}

\item{top_n}{How many of the largest CNA (in number of genes) to get.}

\item{bp_tolerance}{How many bp of tolerance to have around feature start/end positions for top_n largest CNVs.}
}
\value{
seurat_obj
}
\description{
Add meta.data about CNAs to a Seurat object from an infercnv_obj
}
