% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_reduction.R
\name{apply_median_filtering}
\alias{apply_median_filtering}
\title{apply_median_filtering}
\usage{
apply_median_filtering(infercnv_obj, window_size = 7,
  on_observations = TRUE, on_references = TRUE)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{window_size}{Size of the window side centered on the data point to filter (default = 7).}

\item{on_observations}{boolean (default=TRUE), run on observations data (tumor cells).}

\item{on_references}{boolean (default=TRUE), run on references (normal cells).}
}
\value{
infercnv_obj with median filtering applied to observations
}
\description{
Apply a median filtering to the expression matrix within each tumor bounds
}
\examples{
data(data)
data(annots)
data(genes)

infercnv_obj <- infercnv::CreateInfercnvObject(raw_counts_matrix=data, 
                                               gene_order_file=genes,
                                               annotations_file=annots,
                                               ref_group_names=c("normal"))

infercnv_obj <- infercnv::run(infercnv_obj,
                              cutoff=1,
                              out_dir="../example_output", 
                              cluster_by_groups=TRUE, 
                              denoise=TRUE,
                              HMM=FALSE,
                              num_threads=2,
                              no_plot=TRUE)

infercnv_obj <- infercnv::apply_median_filtering(infercnv_obj)
# plot result object

}
