% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_reduction.R
\name{apply_median_filtering}
\alias{apply_median_filtering}
\title{apply_median_filtering}
\usage{
apply_median_filtering(infercnv_obj, window_size = 7,
  on_observations = TRUE, on_references = TRUE)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{window_size}{Size of the window side centered on the data point to filter (default = 7).}

\item{on_observations}{boolean (default=TRUE), run on observations data (tumor cells).}

\item{on_references}{boolean (default=TRUE), run on references (normal cells).}
}
\value{
infercnv_obj with median filtering applied to observations
}
\description{
Apply a median filtering to the expression matrix within each tumor bounds
}
