% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{cellGene}
\alias{cellGene}
\alias{cellGene,MCMC_inferCNV-method}
\title{Access the values for cellGene}
\usage{
cellGene(obj)

\S4method{cellGene}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
A list.
}
\description{
This function returns the list of values in cellGene
}
