% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{clear_noise}
\alias{clear_noise}
\title{clear_noise()}
\usage{
clear_noise(infercnv_obj, threshold)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{threshold}{values within reference mean +- threshold are set to zero.}
}
\value{
infercnv_obj
}
\description{
Remove values that are too close to the reference cell expr average and are considered noise.
}
