% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{collapse_snp2gene}
\alias{collapse_snp2gene}
\title{collapse_snp2gene}
\usage{
collapse_snp2gene(
  infercnv_allele_obj,
  gene_annot,
  collapse_method = c("highest", "median", "mean")
)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele based obj}

\item{gene_annot}{a gene order annotation}

\item{collapse_method}{The method used to collapse snp into gene. Take the highest coverage/median/mean
value within each gene as its representative. default = "highest"}
}
\description{
This function aims to collapse infercnv_allele obj from snp level into gene level
by only leveraging different characteristics
}
