% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{color.palette}
\alias{color.palette}
\title{Helper function allowing greater control over the steps in a color palette.}
\usage{
color.palette(steps, between = NULL, ...)
}
\arguments{
\item{steps}{Vector of colors to change use in the palette}

\item{between}{Steps where gradients change}

\item{...}{Additional arguments of colorRampPalette}
}
\value{
Color palette
}
\description{
Helper function allowing greater control over the steps in a color palette.
             Source: http://menugget.blogspot.com/2011/11/define-color-steps-for-
             colorramppalette.html#more
}
\examples{
color.palette(c("darkblue", "white", "darkred"),
              c(2, 2))

}
