% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextGenHeatMap.R
\name{Create_NGCHM}
\alias{Create_NGCHM}
\title{Create Next Generation Clustered Heat Map (NG-CHM)}
\usage{
Create_NGCHM(plot_data, reference_idx, ref_index, ref_groups, location_data,
  path_to_shaidyMapGen, out_dir, contigs, title, gene_symbol = NULL)
}
\arguments{
\item{plot_data}{Expression data}

\item{reference_idx}{(vector) vector of all reference cell line ID's}

\item{ref_index}{(list) list of index values for each reference group}

\item{ref_groups}{(list) list of cell ID's for each reference groups}

\item{location_data}{(data frame) Data frame with position information for each gene}

\item{path_to_shaidyMapGen}{(string) Path to the java application ShaidyMapGen.jar}

\item{out_dir}{(string) Path to where the infercnv.ngchm output file should be saved to}

\item{contigs}{(list) List of the chr in order}

\item{title}{(string) Title that will be used for the heatmap}

\item{gene_symbal}{(string) Specify the label type that is given to the gene needed to create linkouts, default is NULL}
}
\value{
Exports a NGCHM file named infercnv.ngchm and saves it to the output directory given to infercnv.
}
\description{
Create highly interactive heat maps for single cell expression data using 
Next Generation Clustered Heat Map (NG-CHM). NG-CHM was developed and 
maintained by MD Anderson Department of Bioinformatics and Computational 
Biology in collaboration with In Silico Solutions.
}
