% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_mask_non_DE.R
\name{get_DE_genes_basic}
\alias{get_DE_genes_basic}
\title{get_DE_genes_basic

Retrieves genes identified as significantly DE based on the corresponding statistical test as applied
to the infercnv_obj@expr.data with tumor/normal comparison}
\usage{
get_DE_genes_basic(infercnv_obj, p_val_thresh = 0.05, test.use = "t")
}
\arguments{
\item{p_val_thresh}{maximum p-value for defining statistical signficance}

\item{test.use}{statistical test to use. (default: "t" for t-test)}

\item{return}{all_DE_results[[condition_pair]] = list(tumor=tumor_type,
normal=normal_type,
pvals=pvals,
de_genes=genes)}
}
