% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{get_reference_grouped_cell_indices}
\alias{get_reference_grouped_cell_indices}
\title{get_reference_grouped_cell_indices()

Retrieves the matrix indices for the columns correspoinding to the reference cells.}
\usage{
get_reference_grouped_cell_indices(infercnv_obj)
}
\arguments{
\item{infercnv_obj}{}
}
\value{
vector of column indices
}
