% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{infer_cnv}
\alias{infer_cnv}
\title{Infer CNV changes given a matrix of RNASeq counts.
Output a pdf and matrix of final values.}
\usage{
infer_cnv(data, gene_order, cutoff, reference_obs, transform_data,
  window_length, max_centered_threshold, noise_threshold, num_ref_groups,
  num_obs_groups, out_path, plot_steps = FALSE, contig_tail = (window_length
  - 1)/2, cluster_reference = NULL, method_bound_vis = NA,
  lower_bound_vis = NA, upper_bound_vis = NA)
}
\description{
Args:
   @param data: Expression matrix (genes X samples),
                assumed to be log2(TPM+1) .
   @param gene_order: Ordering of the genes (data's rows) 
                      according to their genomic location
                      To include all genes use 0.
   @param cutoff: Cut-off for the average expression of genes to be 
                  used for CNV inference.
   @param reference_obs: Column names of the subset of samples (data's columns)
                         that should be used as references.
                         If not given, the average of all samples will 
                         be the reference.
   @param transform_data: Indicator to log2 + 1 transform
   @param window_length: Length of the window for the moving average
                         (smoothing). Should be an odd integer.
   @param max_centered_threshold: The maximum value a a value can have after
                                  centering. Also sets a lower bound of
                                  -1 * this value.
   @param noise_threshold: The minimum difference a value can be from the 
                           average reference in order for it not to be
                           removed as noise.
   @param num_ref_groups: The number of reference groups of a list of
                          indicies for each group of reference indices in
                          relation to reference_obs.
   @param num_obs_groups: Number of groups to break the observations into.
   @param pdf_path: The path to what to save the pdf as. The raw data is
                    also written to this path but with the extension .txt .
   @param plot_steps: If true turns on plotting intermediate steps.
   @param contig_tail: Length of the tail removed from the ends of contigs.
   @param cluster_reference: If given, clustering of observation will only be
                             relative to genes on this contig.
   @param method_bound: Method to use for bounding values in the visualization.
   @param lower_bound_vis: Lower bound to normalize data to for visualization.
   @param upper_bound_vis: Upper bound to normalize data to for visualization.
}
\details{
Returns:
   @return: No return.
}

