% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele_BayesNet.R
\name{inferCNVAlleleBayesNet}
\alias{inferCNVAlleleBayesNet}
\title{inferCNVAlleleBayesNet: Run Bayesian NetworkModel To Obtain Posterior Probabilities For HMM Predicted States}
\usage{
inferCNVAlleleBayesNet(infercnv_allele_obj, infercnv_allele_hmm)
}
\description{
Run MCMC method to get posterior probability of cells having Deletions/LOHs
}
