% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\name{inferCNVBayesNet}
\alias{inferCNVBayesNet}
\title{inferCNVBayesNet: Run Bayesian Network Mixture Model To Obtain Posterior Probabilities For HMM Predicted States}
\usage{
inferCNVBayesNet(file_dir, infercnv_obj, HMM_obj,
  model_file = system.file("BUGS_Mixture_Model", package = "infercnv"),
  CORES = 1, out_dir, postMcmcMethod = NULL, plotingProbs = TRUE,
  quietly = TRUE)
}
\arguments{
\item{file_dir}{Location of the directory of the inferCNV outputs.}

\item{infercnv_obj}{InferCNV object.}

\item{HMM_obj}{InferCNV object with HMM states in expression data.}

\item{model_file}{Path to the BUGS Model file.}

\item{CORES}{Option to run parallel by specifying the number of cores to be used.}

\item{out_dir}{(string) Path to where the output file should be saved to.}

\item{postMcmcMethod}{What actions to take after finishing the MCMC.}

\item{plotingProbs}{Option for adding plots of Cell and CNV probabilities.}

\item{quietly}{Option to print descriptions along each step.}
}
\value{
Returns a MCMC_inferCNV_obj and posterior probability of being in one of six Copy Number Variation states 
(states: 0, 0.5, 1, 1.5, 2, 3) for CNV's identified by inferCNV's HMM.
}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior 
probability of being in one of six Copy Number Variation states (states: 0, 0.5, 1, 1.5, 2, 3) for CNV's identified by 
inferCNV's HMM. Posterior probabilities are found for the entire CNV cluster and each individual
cell line in the CNV.
}
