% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\name{inferCNVBayesNet}
\alias{inferCNVBayesNet}
\title{inferCNVBayesNet: Run Bayesian Network Mixture Model To Obtain Posterior Probabilities For HMM Predicted States}
\usage{
inferCNVBayesNet(file_dir, model_file = system.file("BUGS_Mixture_Model",
  package = "infercnv"), CORES = NULL, output_dir,
  postMcmcMethod = NULL, plotingProbs = TRUE)
}
\arguments{
\item{infercnv_dir}{Location of the directory of the inferCNV outputs.}

\item{model}{Path to the BUGS Model file.}

\item{parallel}{Option to run parallel by specifying the number of cores to be used.}

\item{out_dir}{(string) Path to where the output file should be saved to.}

\item{method}{What actions to take after finishing the MCMC.}
}
\value{
Returns a MCMC_inferCNV_obj and posterior probability of being in one of six Copy Number Variation states 
(states: 0, 0.5, 1, 1.5, 2, 3) for CNV's identified by inferCNV's HMM.
}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior 
probability of being in one of six Copy Number Variation states (states: 0, 0.5, 1, 1.5, 2, 3) for CNV's identified by 
inferCNV's HMM. Posterior probabilities are found for the entire CNV cluster and each individual
cell line in the CNV.
}
