% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\name{inferCNVBayesNet}
\alias{inferCNVBayesNet}
\title{inferCNVBayesNet: Run Bayesian Network Mixture Model To Obtain Posterior Probabilities For HMM Predicted States}
\usage{
inferCNVBayesNet(file_dir, infercnv_obj, HMM_obj, out_dir,
  model_file = system.file("BUGS_Mixture_Model", package = "infercnv"),
  CORES = 1, postMcmcMethod = NULL, plotingProbs = TRUE,
  quietly = TRUE, diagnostics = FALSE)
}
\arguments{
\item{file_dir}{Location of the directory of the inferCNV outputs.}

\item{infercnv_obj}{InferCNV object.}

\item{HMM_obj}{InferCNV object with HMM states in expression data.}

\item{out_dir}{(string) Path to where the output file should be saved to.}

\item{model_file}{Path to the BUGS Model file.}

\item{CORES}{Option to run parallel by specifying the number of cores to be used. (Default: 1)}

\item{postMcmcMethod}{What actions to take after finishing the MCMC.}

\item{plotingProbs}{Option for adding plots of Cell and CNV probabilities. (Default: TRUE)}

\item{quietly}{Option to print descriptions along each step. (Default: TRUE)}

\item{diagnostics}{Option to plot Diagnostic plots and tables. (Default: FALSE)}
}
\value{
Returns a MCMC_inferCNV_obj and posterior probability of being in one of six Copy Number Variation states
(states: 0, 0.5, 1, 1.5, 2, 3) for CNV's identified by inferCNV's HMM.
}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior
probability of being in one of six Copy Number Variation states (states: 0, 0.5, 1, 1.5, 2, 3) for CNV's identified by
inferCNV's HMM. Posterior probabilities are found for the entire CNV cluster and each individual
cell line in the CNV.
}
\examples{
data(data)
data(annots)
data(genes)

infercnv_obj <- infercnv::CreateInfercnvObject(raw_counts_matrix=data, 
                                               gene_order_file=genes,
                                               annotations_file=annots,
                                               ref_group_names=c("normal"))
infercnv_obj <- infercnv::run(infercnv_obj,
                              cutoff=1,
                              out_dir="../example_output", 
                              cluster_by_groups=TRUE, 
                              denoise=TRUE,
                              HMM=TRUE,
                              num_threads=2,
                              no_plot=TRUE)

files <- list.files("../example_output", full.names = TRUE)
HMM_obj <- readRDS(files[grep("hmm_mode-samples.infercnv_obj",files)])
mcmc_obj <- infercnv::inferCNVBayesNet( infercnv_obj   = infercnv_obj,
                              HMM_obj         = HMM_obj,
                              file_dir        = "../example_output",
                              postMcmcMethod  = "removeCNV",
                              out_dir         = "../example_output",
                              quietly         = TRUE,
                              CORES           = 2,
                              plotingProbs    = FALSE,
                              diagnostics     = FALSE)
}
