% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_combined_BayesNet.R
\name{inferCNVCombinedBayesNet}
\alias{inferCNVCombinedBayesNet}
\title{inferCNVCombinedBayesNet: Run Bayesian Network Mixture Model Leveraging both
Expression and Allele Data To Obtain Posterior Probabilities For HMM Predicted States}
\usage{
inferCNVCombinedBayesNet(
  combined_file_path,
  combined_file_token,
  infercnv_obj,
  infercnv_allele_obj,
  allele_mode = c("snp_level", "gene_level"),
  method = c("common", "union"),
  type = c("i6", "i3"),
  enable_cnLOH = TRUE,
  output_path,
  cores = 5
)
}
\arguments{
\item{combined_file_path}{Location of the directory of the combined HMM outputs.}

\item{combined_file_token}{(string) String token that contains some info on settings used to name combined-based files.}

\item{infercnv_obj}{InferCNV object.}

\item{infercnv_allele_obj}{InferCNV allele object.}

\item{allele_mode}{The type of allele data provided. "snp_level" or "gene_level".}

\item{method}{The way integrating HMM boundaries from two methods: union/common.
"union" keeps those expression-specific gene that are not found in allele data during modeling,
while "common" removes those ones.}

\item{type}{The type of combined model running in Bayesian Method. i6 or i3.}

\item{enable_cnLOH}{Option for detecting cnLOH events by adding one more state in Bayesian Method. (Default: True)}

\item{output_path}{(string) Path to where the output file should be saved to.}

\item{cores}{Option to run parallel by specifying the number of cores to be used. (Default: 5)}
}
\value{
Returns a MCMC_infercnv_combined obj and posterior probability of being in one of three/six Copy Number Variation states
(i6 states: 1-2,Deletion; 3,Neutral; 4-6,Amplification; 7,cnLOH,if enable_cnLOH),
(i3 states: 1,Deletion; 2,Neutral; 3,Amplification; 4,cnLOH,if enable_cnLOH), for CNV's identified by inferCNV's HMM.
}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior
probability of being in one of three/six Copy Number Variation states 
(i6 states: 1-2,Deletion; 3,Neutral; 4-6,Amplification; 7,cnLOH,if enable_cnLOH),
(i3 states: 1,Deletion; 2,Neutral; 3,Amplification; 4,cnLOH,if enable_cnLOH),
for CNV's identified by inferCNV's HMM. Posterior probabilities are found for the entire CNV cluster 
and each individual cell line in the CNV.
}
\examples{
data(infercnv_data_example)
data(infercnv_annots_example)
data(infercnv_genes_example)
data(infercnv_allele_alt_example)
data(infercnv_allele_tot_example)
data(infercnv_object_allele_gene_example)

infercnv_object_example <- infercnv::CreateInfercnvObject(raw_counts_matrix = infercnv_data_example,
                                                          raw_allele_matrix = infercnv_allele_alt_example,
                                                          raw_coverage_matrix = infercnv_allele_tot_example,
                                                          gene_order_file = infercnv_genes_example,
                                                          annotations_file = infercnv_annots_example,
                                                          ref_group_names=c("Microglia/Macrophage"),
                                                          snp_split_by = "::")
          
out_dir = tempfile()
file_gene_token <- "HMM_gene_pred"
infercnv_object_example <- infercnv::run(infercnv_object_example,
                                         cutoff=1,
                                         out_dir=out_dir, 
                                         cluster_by_groups=F, 
                                         denoise=FALSE,
                                         HMM=TRUE,
                                         analysis_mode = "samples",
                                         no_plot=TRUE)
                                         
hmm_allele_gene_obj_HMM_samples <- infercnv:::allele_HMM_predict_CNV_via_HMM_on_whole_tumor_samples(infercnv_object_allele_gene_example,
                                                                                                    trim = 0)
infercnv:::generate_cnv_region_reports(hmm_allele_gene_obj_HMM_samples, 
                                       output_filename_prefix=file_gene_token,
                                       out_dir=out_dir,
                                       ignore_neutral_state = 2,
                                       by="subcluster")
                                       
combined_obj <- infercnv::fusion_HMM_report_two_methods(expression_file_path = out_dir,
                                                        expression_file_token = "17_HMM_predHMMi6.hmm_mode-samples",
                                                        allele_file_path = out_dir,
                                                        allele_file_token = file_gene_token,
                                                        infercnv_obj = infercnv_object_example,
                                                        infercnv_allele_obj = infercnv_object_allele_gene_example,
                                                        method = "common",
                                                        type = "i6",
                                                        cluster_by_groups=F,
                                                        output_path = out_dir,
                                                        output_prefix = "combined")
mcmc_combined <- infercnv::inferCNVCombinedBayesNet(combined_file_path = out_dir,
                                                    combined_file_token = "combined",
                                                    infercnv_obj = infercnv_object_example,
                                                    infercnv_allele_obj = infercnv_object_allele_gene_example,
                                                    allele_mode = "gene_level",
                                                    method = "common",
                                                    type = "i6",
                                                    enable_cnLOH = TRUE,
                                                    output_path = file.path(out_dir,"bayesian_folder"))
}
