% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{map2gene}
\alias{map2gene}
\title{map2gene}
\usage{
map2gene(infercnv_allele_obj, gene_order)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele based obj}

\item{gene_order}{gene annotation file}
}
\description{
This function aims to map snp data back to gene data given the annotation file
}
