% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{map2gene}
\alias{map2gene}
\title{map2gene}
\usage{
map2gene(infercnv_allele_obj)
}
\arguments{
\item{infercnv_allele_obj}{infercnv object}
}
\value{
infercnv object
}
\description{
This function aims to map snp-based data back to gene-based data given the annotation file provided
}
\examples{
data(infercnv_object_example)

infercnv_object_allele_example <- infercnv::setAlleleMatrix(infercnv_object_example@.allele ,
                                                            snp_min_coverage = 0,
                                                            snp_min.cell = 2)
infercnv_object_allele_example <- infercnv::map2gene(infercnv_object)

}
