% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_mask_non_DE.R
\name{mask_non_DE_genes_basic}
\alias{mask_non_DE_genes_basic}
\title{mask_non_DE_genes_basic()}
\usage{
mask_non_DE_genes_basic(infercnv_obj, p_val_thresh = 0.05,
  test.use = "wilcoxon", center_val = mean(infercnv_obj@expr.data))
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{p_val_thresh}{p-value threshold for assigning DE statistically significant}

\item{test.use}{statistical test to use.  (default: "wilcoxon") alternatives include 'perm' or 't'.}

\item{center_val}{value to assign to those genes that are not found to be statistically DE.}
}
\value{
infercnv_obj
}
\description{
Mask gene expression in infercnv_obj based on expression values that are found
statistically significantly different between tumor/normal based on specified
statistical test.

Uses the expression data stored in infercnv_obj@expr.data
}
