% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele_BayesNet.R
\docType{class}
\name{MCMC_infercnv_allele-class}
\alias{MCMC_infercnv_allele-class}
\alias{MCMC_infercnv_allele}
\title{MCMC infercnv_allele class -- under development}
\description{
This class extends the functionality of infercnv_allele class
aiming to coorporate MCMC object leveraging Bayesian model

Slots in the MCMC_infercnv_allele object (besides infercnv) include:
}
\section{Slots}{

\describe{
\item{\code{bugs_model}}{a string path to where the bug file locates}

\item{\code{n.array}}{the list of arrays containing allele count/ ref count etc.}

\item{\code{n.meta}}{metadata info}

\item{\code{posterior_prob}}{the list obj containing posterior prob from bayesian model}
}}

