% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele_BayesNet.R
\docType{class}
\name{MCMC_infercnv_allele-class}
\alias{MCMC_infercnv_allele-class}
\alias{MCMC_infercnv_allele}
\title{MCMC infercnv_allele class}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior
probability of being in one of two Copy Number Variation states 
(i2 states: 1,Deletion; 2,Neutral) for CNV's identified by inferCNV_allele's HMM.
}
\section{Slots}{

\describe{
\item{\code{bugs_model}}{BUGS model.}

\item{\code{cell_gene}}{List containing the Cells and Genes that make up each CNV.}

\item{\code{cnv_probabilities}}{Probabilities of each CNV belonging to a particular state from 0 (least likely) to 1 (most likely).}

\item{\code{cell_probabilities}}{Probabilities of each cell being in a particular state, from 0 (least likely) to 1 (most likely).}

\item{\code{cnv_regions}}{ID for each CNV found by the HMM.}
}}

