% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_combined_BayesNet.R
\docType{class}
\name{MCMC_infercnv_combined-class}
\alias{MCMC_infercnv_combined-class}
\alias{MCMC_infercnv_combined}
\title{MCMC infercnv_combined class}
\description{
Uses Markov Chain Monte Carlo (MCMC) and Gibbs sampling to estimate the posterior
probability of being in one of three/six Copy Number Variation states 
(i6 states: 1-2,Deletion; 3,Neutral; 4-6,Amplification; 7,cnLOH,if enable_cnLOH),
(i3 states: 1,Deletion; 2,Neutral; 3,Amplification; 4,cnLOH,if enable_cnLOH),
for CNV's identified by inferCNV's HMM. Posterior probabilities are found for the entire CNV cluster 
and each individual cell line in the CNV.
}
\section{Slots}{

\describe{
\item{\code{bugs_model}}{BUGS model.}

\item{\code{cell_gene}}{List containing the Cells and Genes that make up each CNV.}

\item{\code{cnv_probabilities}}{Probabilities of each CNV belonging to a particular state from 0 (least likely)to 1 (most likely).}

\item{\code{cell_probabilities}}{Probabilities of each cell being in a particular state, from 0 (least likely)to 1 (most likely).}

\item{\code{cnv_regions}}{ID for each CNV found by the HMM.}

\item{\code{infercnv_obj}}{InferCNV obj.}

\item{\code{infercnv_obj_mu}}{Mean values to be used for determining the distribution of each cell line.}

\item{\code{infercnv_obj_sig}}{fitted values for cell lines, 1/standard deviation to be used for determining the distribution of each cell line.}

\item{\code{infercnv_allele_obj}}{InferCNV_allele obj.}
}}

