% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_combined_BayesNet.R
\docType{class}
\name{MCMC_infercnv_combined-class}
\alias{MCMC_infercnv_combined-class}
\alias{MCMC_infercnv_combined}
\title{MCMC infercnv_combined class}
\description{
This class extends the functionality of infercnv/infercnv_allele class
aiming to cooperate MCMC object leveraging Bayesian model

Slots in the MCMC_infercnv_combined object include:
}
\section{Slots}{

\describe{
\item{\code{bugs_model}}{BUGS model.}

\item{\code{cell_gene}}{List containing the Cells and Genes that make up each CNV.}

\item{\code{cnv_probabilities}}{Probabilities of each CNV belonging to a particular state from 0 (least likely)to 1 (most likely).}

\item{\code{cell_probabilities}}{Probabilities of each cell being in a particular state, from 0 (least likely)to 1 (most likely).}

\item{\code{cnv_regions}}{ID for each CNV found by the HMM.}

\item{\code{infercnv_obj}}{InferCNV obj.}

\item{\code{infercnv_obj_mu}}{Mean values to be used for determining the distribution of each cell line.}

\item{\code{infercnv_obj_sig}}{fitted values for cell lines, 1/standard deviation to be used for determining the distribution of each cell line.}

\item{\code{infercnv_allele_obj}}{InferCNV_allele obj.}

\item{\code{States}}{NOT defined yet.}
}}

