% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{nonParallel}
\alias{nonParallel}
\alias{nonParallel,MCMC_inferCNV-method}
\title{Run simulations in Non-Parallel mode}
\usage{
nonParallel(obj)

\S4method{nonParallel}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Run simulations in Non-Parallel mode
}
