% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV.R
\name{order_reduce}
\alias{order_reduce}
\title{Order the data and subset the data to data in the genomic position file.}
\usage{
order_reduce(data, genomic_position)
}
\arguments{
\item{data}{Data (expression) matrix where the row names should be in
the row names of the genomic_position file.}

\item{genomic_position}{Data frame read in from the genomic position file}
}
\value{
Returns a matrix of expression in the order of the
           genomic_position file. NULL is returned if the genes in both
           data parameters do not match.
}
\description{
Args:
}
