% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{removeCNV}
\alias{removeCNV}
\alias{removeCNV,MCMC_inferCNV-method}
\title{Run simulations and remove CNV's that have a probability of being normal above a set thresholld. 
This removes possible false posotives identified by the HMM.}
\usage{
removeCNV(obj)

\S4method{removeCNV}{MCMC_inferCNV}(obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
obj The MCMC_inferCNV_obj S4 object.
}
\description{
Run simulations and remove CNV's that have a probability of being normal above a set thresholld. 
This removes possible false posotives identified by the HMM.
}
