% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\name{returningInferCNV}
\alias{returningInferCNV}
\title{Return the InferCNV Object with the new adjucted CNV's}
\usage{
returningInferCNV(obj, infercnv_obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}
}
\value{
An inferCNV object
}
\description{
Returns Infercnv Object
}
