% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{returningInferCNV}
\alias{returningInferCNV}
\alias{returningInferCNV,MCMC_inferCNV-method}
\title{Return the InferCNV Object with the new adjucted CNV's}
\usage{
returningInferCNV(obj, infercnv_obj)

\S4method{returningInferCNV}{MCMC_inferCNV}(obj, infercnv_obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}

\item{infercnv_obj}{Current inferCNV object that will be adjusted based on the results of the Bayesian Network Model.}
}
\value{
An inferCNV object
}
\description{
Returns Infercnv Object
}
\keyword{internal}
