% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{returningInferCNV}
\alias{returningInferCNV}
\alias{returningInferCNV,MCMC_inferCNV-method}
\title{Return the InferCNV Object with the new adjucted CNV's}
\usage{
returningInferCNV(obj, infercnv_obj)

\S4method{returningInferCNV}{MCMC_inferCNV}(obj, infercnv_obj)
}
\arguments{
\item{obj}{The MCMC_inferCNV_obj S4 object.}

\item{infercnv_obj}{Current inferCNV object that will be adjusted based on the results of the Bayesian Network Model.}
}
\value{
An inferCNV object
}
\description{
Returns Infercnv Object
}
\examples{
# data(data)
# data(annots)
# data(genes)
#
# infercnv_obj <- infercnv::CreateInfercnvObject(raw_counts_matrix=data, 
#                                                gene_order_file=genes,
#                                                annotations_file=annots,
#                                                ref_group_names=c("normal"))
# infercnv_obj <- infercnv::run(infercnv_obj,
#                               cutoff=1,
#                               out_dir="../example_output", 
#                               cluster_by_groups=TRUE, 
#                               denoise=TRUE,
#                               HMM=TRUE,
#                               num_threads=2,
#                               no_plot=TRUE)
# files <- list.files("../example_output", full.names = TRUE)
# HMM_obj <- readRDS(files[grep("hmm_mode-samples.infercnv_obj",files)])
# mcmc_obj <- infercnv::inferCNVBayesNet( infercnv_obj   = infercnv_obj,
#                               HMM_obj         = HMM_obj,
#                               file_dir        = "../example_output",
#                               postMcmcMethod  = "removeCNV",
#                               out_dir         = "../example_output",
#                               quietly         = TRUE,
#                               CORES           = 2,
#                               plotingProbs    = FALSE,
#                               diagnostics     = FALSE)

data(HMM_obj)
data(mcmc_obj)

hmm.infercnv_obj <- infercnv::returningInferCNV(mcmc_obj, HMM_obj)


}
\keyword{internal}
