% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_BayesNet.R
\docType{methods}
\name{returningInferCNV}
\alias{returningInferCNV}
\alias{returningInferCNV,MCMC_inferCNV-method}
\title{Return the InferCNV Object with the new adjucted CNV's}
\usage{
returningInferCNV(infercnv_obj, HMM_states)

\S4method{returningInferCNV}{MCMC_inferCNV}(infercnv_obj, HMM_states)
}
\arguments{
\item{infercnv_obj}{Current inferCNV object that will be adjusted based on the results of the Bayesian Network Model.}

\item{HMM_states}{InferCNV object with HMM states in expression data.}
}
\value{
An inferCNV object
}
\description{
Returns Infercnv Object
}
\examples{
data(HMM_states)
data(mcmc_obj)

hmm.infercnv_obj <- infercnv::returningInferCNV(mcmc_obj, HMM_states)


}
\keyword{internal}
