% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{run}
\alias{run}
\title{run() : Invokes a routine inferCNV analysis to Infer CNV changes given a matrix of RNASeq counts.}
\usage{
run(infercnv_obj, cutoff = 1, min_cells_per_gene = 3, out_dir = ".",
  normalize_factor = NA, window_length = 101, num_ref_groups = NULL,
  max_centered_threshold = NA, noise_filter = NA, sd_amplifier = 1.5,
  cluster_by_groups = FALSE, k_obs_groups = 1,
  outlier_method_bound = "average_bound", outlier_lower_bound = NA,
  outlier_upper_bound = NA, hclust_method = "complete",
  anscombe_normalize = TRUE, use_zscores = FALSE,
  remove_genes_at_chr_ends = FALSE, mask_nonDE_genes = TRUE,
  mask_nonDE_pval = 0.05, test.use = "wilcoxon", plot_steps = FALSE,
  debug = FALSE, include.spike = FALSE, pseudocount = 0)
}
\arguments{
\item{infercnv_obj}{An infercnv object populated with raw count data}

\item{cutoff}{Cut-off for the min average read counts per gene among reference cells. (default: 1)}

\item{min_cells_per_gene}{minimum number of reference cells requiring expression measurements to include the corresponding gene. 
default: 3}

\item{out_dir}{path to directory to deposit outputs (default: '.')}

\item{normalize_factor}{scaling factor for total sum of counts (default: NA, in which case
will be set = 10^round(log10(mean(colSums))), typically setting to 1e5}

\item{window_length}{Length of the window for the moving average
(smoothing). Should be an odd integer. (default: 101)#'}

\item{num_ref_groups}{The number of reference groups or a list of
indices for each group of reference indices in
relation to reference_obs. (default: NULL)}

\item{max_centered_threshold}{The maximum value a value can have after
centering. Also sets a lower bound of
-1 * this value.}

\item{noise_filter}{Values +- from the reference cell mean will be set to zero (whitening effect)
default(NA, instead will use sd_amplifier below.}

\item{sd_amplifier}{Noise is defined as mean(reference_cells) +- sdev(reference_cells) * sd_amplifier
default: 1.0}

\item{cluster_by_groups}{If observations are defined according to groups (ie. patients), each group
of cells will be clustered separately. (default=FALSE, instead will use k_obs_groups setting)}

\item{k_obs_groups}{Number of groups in which to break the observations. (default: 1)}

\item{outlier_method_bound}{Method to use for bounding outlier values. (default: "average_bound")
Will preferentially use outlier_lower_bounda and outlier_upper_bound if set.}

\item{outlier_lower_bound}{Outliers below this lower bound will be set to this value.}

\item{outlier_upper_bound}{Outliers above this upper bound will be set to this value.}

\item{hclust_method}{Method used for hierarchical clustering of cells. Valid choices are:
"ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid".}

\item{anscombe_normalize}{Perform anscombe normalization on normalized counts before log transformation.}

\item{use_zscores}{If true, converts log(expression) data to zscores based on reference cell expr distribution.}

\item{remove_genes_at_chr_ends}{If true, removes the window_length/2 genes at both ends of the chromosome.}

\item{mask_nonDE_genes}{If true, sets genes not significantly differentially expressed between tumor/normal to
the mean value for the complete data set}

\item{mask_nonDE_pval}{p-value threshold for defining statistically significant DE genes between tumor/normal}

\item{plot_steps}{If true, saves infercnv objects and plots data at the intermediate steps.}
}
\value{
infercnv_obj containing filtered and transformed data
}
\description{
Function doing the actual analysis before calling the plotting functions.
}
