% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{setAlleleMatrix}
\alias{setAlleleMatrix}
\title{setAlleleMatrix}
\usage{
setAlleleMatrix(
  infercnv_allele_obj,
  snp_min_coverage = 0,
  min_cells_per_snp = 3
)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele_object}

\item{snp_min_coverage}{the minimum number of threshold that each instance should have in allele/coverage data. 
Each instance that is less than this number will be replaced with 0. default = 0 (No action)}

\item{snp_filter}{a boolean value whether to do pre-filtering for allele matrix. default = TRUE}

\item{snp_min.cell}{a threshold used to filter out non-heterozygous snps. The minimum number of (normal) cells 
having both alleles. default = 3}
}
\value{
infercnv_allele
}
\description{
This function initializes the lesser allele fraction/count determined by at least the minimum
number of snps having allele fraction less/more than 0.5 with the minimum of cells supported. 
lesser allele fraction will be stored in the slot @expr.data. lesser allele count 
will be stored in the slot @count.data
}
\examples{
data(infercnv_object_example)

infercnv_object_allele_example <- infercnv::setAlleleMatrix(infercnv_object_example@.allele ,
                                                            snp_min_coverage = 0,
                                                            snp_min.cell = 2)

}
