% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_allele.R
\name{setAlleleMatrix_HB}
\alias{setAlleleMatrix_HB}
\title{setAlleleMatrix_HB}
\usage{
setAlleleMatrix_HB(
  infercnv_allele_obj,
  snp_min_coverage = 2,
  snp_filter = TRUE,
  snp_het.deviance.threshold = 0.1,
  snp_min.cell = 3,
  smooth_method = "runmeans",
  window_length = 101
)
}
\arguments{
\item{infercnv_allele_obj}{infercnv_allele_object}

\item{snp_min_coverage}{the minimum number of counts that the snp express in coverage data. default = 2}

\item{snp_filter}{a boolean value whether to do pre-filtering for allele matrix. default = TRUE}

\item{snp_het.deviance.threshold}{a threshold used to filter 
out non-hetergozous snps. default = 0.1}

\item{snp_min.cell}{a threshold used to filter out snps that express less than 
the minimum number of cells. default = 3}

\item{smooth_method}{a method used for smoothing allele fraction default = "runmeans"}

\item{window_length}{smoothing window size. default = 101}
}
\description{
This function mimics the way the HB does aiming to initialize the lesser allele fraction/count.
lesser allele fraction will be stored in the slot @expr.data. lesser allele count 
will be stored in the slot @count.data
}
