% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_mask_non_DE.R
\name{seurat_from_infercnv_obj}
\alias{seurat_from_infercnv_obj}
\title{seurat_from_infercnv_obj()

Generate a Seurat object from an infercnv_obj}
\usage{
seurat_from_infercnv_obj(infercnv_obj, normalize_factor)
}
\arguments{
\item{infercnv_obj}{}

\item{normalize_factor}{}
}
