% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{split_references}
\alias{split_references}
\title{split_references()}
\usage{
split_references(infercnv_obj, num_groups = 2, hclust_method = "complete")
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{num_groups}{(default: 2)}

\item{hclust_method}{clustering method to use (default: 'complete')}
}
\value{
infercnv_obj
}
\description{
Split up reference observations in to k groups based on hierarchical clustering.
}
