% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_ops.R
\name{subtract_ref_expr_from_obs}
\alias{subtract_ref_expr_from_obs}
\title{subtract_ref_expr_from_obs()}
\usage{
subtract_ref_expr_from_obs(infercnv_obj, inv_log = FALSE,
  use_bounds = TRUE)
}
\arguments{
\item{infercnv_obj}{infercnv_object}

\item{inv_log}{mean values will be determined based on (2^x -1)}

\item{use_bounds}{if multiple normal data sets are used, it takes the bounds of the means from each set for subtraction.
Alternatively, will use the mean( mean(normal) for each normal)    default: TRUE}
}
\value{
infercnv_obj containing the reference subtracted values.
}
\description{
Remove the average of the genes of the reference observations from all
observations' expression. In the case there are multiple reference groupings,
the averages are computed separately for each reference grouping, and the min|max
of the averages are subtracted from the observation expression levels. Any values within the range
of the min,max of the group are set to zero.
}
\details{
Subtracting the mean of the reference expr distributions from the observed cells.
}
